### 2020-11-10 ac9449b
FindHDF5: remove dupe variable
