+++ .mypy.ini
@@ -1,4 +1,6 @@
 [mypy]
+files = scripts/
+
 ignore_missing_imports = True
 strict_optional = False
 allow_redefinition = True
