+++ .github/workflows/ci_mac_meson.yml
@@ -1,37 +1,0 @@
-name: ci_mac_meson
-
-on:
-  push:
-    paths:
-      - "**/meson.build"
-      - ".github/workflows/ci_mac_meson.yml"
-
-jobs:
-
-  meson:
-    if: false  # some glitch with CI meson, not apparent on physical Mac
-    runs-on: macos-latest
-    steps:
-    - uses: actions/checkout@v2
-    - uses: actions/setup-python@v2
-      with:
-        python-version: '3.x'
-
-    - run: pip install meson
-    - run: brew install gcc hdf5 ninja
-
-    - run: meson setup build
-
-    - run: meson compile -C build
-    - uses: actions/upload-artifact@v1
-      if: failure()
-      with:
-        name: Mac_Meson_Buildlog
-        path: build/meson-logs/meson-log.txt
-
-    - run: meson test -C build -v
-    - uses: actions/upload-artifact@v1
-      if: failure()
-      with:
-        name: Mac_Meson_Testlog
-        path: build/meson-logs/testlog.txt
