+++ .github/workflows/ci_linux_meson.yml
@@ -1,38 +1,0 @@
-name: ci_linux_meson
-
-on:
-  push:
-    paths:
-      - "**/meson.build"
-      - ".github/workflows/ci_linux_meson.yml"
-
-
-jobs:
-
-  meson:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v2
-    - uses: actions/setup-python@v2
-      with:
-        python-version: '3.x'
-
-    - run: pip install meson ninja
-    - run: |
-        sudo apt update -yq
-        sudo apt install -yq --no-install-recommends gfortran libhdf5-dev
-
-    - run: meson setup build
-    - run: meson compile -C build
-    - uses: actions/upload-artifact@v1
-      if: failure()
-      with:
-        name: Linux_Meson_Buildlog
-        path: build/meson-logs/meson-log.txt
-
-    - run: meson test -C build -v
-    - uses: actions/upload-artifact@v1
-      if: failure()
-      with:
-        name: Linux_Meson_Testlog
-        path: build/meson-logs/testlog.txt
