+++ .github/workflows/ci_mac.yml
@@ -1,34 +1,0 @@
-name: ci_mac
-
-env:
-  FC: gfortran-9
-  CC: gcc-9
-
-on:
-  push:
-    paths:
-      - "**/CMakeLists.txt"
-      - "**.cmake"
-      - ".github/workflows/ci_mac.yml"
-
-jobs:
-
-  default:
-    runs-on: macos-latest
-    steps:
-    - uses: actions/checkout@v2
-
-    - run: ctest -S setup.cmake -VV
-
-  # examples:
-  # works but saving CI time
-  #   runs-on: macos-latest
-  #   steps:
-  #   - uses: actions/checkout@v2
-
-  #   - run: cmake -B Examples/build -S Examples
-  #   - run: cmake --build Examples/build
-
-  #   - name: test examples
-  #     run: ctest -V
-  #     working-directory: Examples/build
