+++ .github/workflows/ci_meson.yml
@@ -1,0 +1,67 @@
+name: ci_meson
+
+on:
+  push:
+    paths:
+      - "**/meson.build"
+      - ".github/workflows/ci_linux_meson.yml"
+
+
+jobs:
+
+  linux:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+    - uses: actions/setup-python@v2
+      with:
+        python-version: '3.x'
+
+    - run: pip install meson
+    - run: |
+        sudo apt update -yq
+        sudo apt install -yq --no-install-recommends gfortran ninja-build libhdf5-dev
+
+    - run: meson setup build
+    - run: meson compile -C build
+    - uses: actions/upload-artifact@v1
+      if: failure()
+      with:
+        name: Linux_Meson_Buildlog
+        path: build/meson-logs/meson-log.txt
+
+    - run: meson test -C build -v
+    - uses: actions/upload-artifact@v1
+      if: failure()
+      with:
+        name: Linux_Meson_Testlog
+        path: build/meson-logs/testlog.txt
+
+  mac:
+    if: false  # some glitch with CI meson, not apparent on physical Mac
+    needs: linux
+    runs-on: macos-latest
+    steps:
+    - uses: actions/checkout@v2
+    - uses: actions/setup-python@v2
+      with:
+        python-version: '3.x'
+
+    - run: pip install meson
+    - run: brew install gcc hdf5 ninja
+
+    - run: meson setup build
+
+    - run: meson compile -C build
+    - uses: actions/upload-artifact@v1
+      if: failure()
+      with:
+        name: Mac_Meson_Buildlog
+        path: build/meson-logs/meson-log.txt
+
+    - run: meson test -C build -v
+    - uses: actions/upload-artifact@v1
+      if: failure()
+      with:
+        name: Mac_Meson_Testlog
+        path: build/meson-logs/testlog.txt
