+++ .github/workflows/ci_linux.yml
@@ -1,33 +1,0 @@
-name: ci_linux
-
-env:
-  FC: gfortran-9
-  CC: gcc-9
-
-on:
-  push:
-    paths:
-      - "**/CMakeLists.txt"
-      - "**.cmake"
-      - "**.f90"
-      - ".github/workflows/ci_linux.yml"
-
-jobs:
-
-  default:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v2
-
-    - run: ctest -S setup.cmake -VV
-
-  examples:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v2
-
-    - run: cmake -B Examples/build -S Examples
-    - run: cmake --build Examples/build
-
-    - run: ctest -V
-      working-directory: Examples/build
