### 2020-11-13 #2820 @kthyng
Hi @scivision! Thanks for your submission. The editorial board is a bit borderline about your submission as to whether it is in scope as research software and is more than a thin wrapper, though everyone thinks it is clearly very useful. However, enough people support it as in scope to move forward with review. So, let's proceed!
