+++ CMakeLists.txt
@@ -5,7 +5,7 @@ include(cmake/compiler_find.cmake)
 
 project(h5fortran
   LANGUAGES C Fortran
-  VERSION 3.4.1
+  VERSION 3.4.2
   DESCRIPTION "thin, light object-oriented HDF5 Fortran interface"
   HOMEPAGE_URL https://github.com/geospace-code/h5fortran)
 
