### 2020-11-23 #2842 @pdebuyl
Apparently, our editorial bot is pinging reviews after a week of inactivity. I usually do that "manually", so here goes for the weekly reminder.
