### 2020-11-25 5b3b7e3
handle missing H5pubconf.h correctly
