+++ setup.cmake
@@ -26,12 +26,11 @@ if(NOT DEFINED CTEST_CMAKE_GENERATOR AND CMAKE_VERSION VERSION_GREATER_EQUAL 3.1
   endif(_gen)
 endif()
 if(NOT DEFINED CTEST_CMAKE_GENERATOR)
+  set(CTEST_BUILD_FLAGS -j)  # not --parallel as this goes to generator directly
   if(WIN32)
     set(CTEST_CMAKE_GENERATOR "MinGW Makefiles")
-    set(CTEST_BUILD_FLAGS -j)  # not --parallel as this goes to generator directly
   else()
     set(CTEST_CMAKE_GENERATOR "Unix Makefiles")
-    set(CTEST_BUILD_FLAGS -j)  # not --parallel as this goes to generator directly
   endif()
 endif()
 
