### 2020-11-25 9b56bd6
add ability to write HDF5 soft links
