### 2020-12-05 geospace-code/h5fortran#9 @bast
On the documentation webpage the button "The Web" points to 404 (outdated URL)
This is top left on https://geospace-code.github.io/h5fortran/

Also the "GitHub" button points to the URL pre-rename but is forwarded.

Stumbled into this while working on https://github.com/openjournals/joss-reviews/issues/2842.
