### 2020-12-05 geospace-code/h5fortran#10 @bast
It may not be clear to the user that the test step also submits a report to CDash
I was following the build instructions at https://github.com/geospace-code/h5fortran#build

These contain a test step (which is very good and worked for me) and the test step contains a CDash step which reports to a public dashboard. The latter is useful for the project and not a problem per se but I think it might be useful to add a note to the documentation (README.md). The general user may not be aware that some information from their computer is being sent to a public dashboard.

(working on https://github.com/openjournals/joss-reviews/issues/2842)
