### 2020-12-05 geospace-code/h5fortran#11 @bast
gfortran compile example in the install instructions does not seem to be enough
I tried "if not using CMake FetchContent as is suggested" under https://github.com/geospace-code/h5fortran/blob/master/Install.md#optional-install and this seems to be lacking libraries.

Later I found this: "On Ubuntu it looks like" https://github.com/geospace-code/h5fortran/tree/master/Examples#non-cmake-build which looks more complete. It seems that the two examples do not match.

(working on https://github.com/openjournals/joss-reviews/issues/2842)
