### 2020-12-05 geospace-code/h5fortran#12 @bast
Documentation about autobuild seems inconsistent with "Examples"
The documentation at https://github.com/geospace-code/h5fortran#autobuild-hdf5 mentions that "h5fortran will automatically build the HDF5 library if needed. This takes a minute or two on a typical laptop."

This was indeed the case when following the example further up.

However, when I later tried https://github.com/geospace-code/h5fortran/tree/master/Examples, CMake configuration told me that "h5fortran: HDF5 not working". This required installing HDF5. Which is not a problem but I expected also here that it would build it for me.

Unless I missed something, the documentation could be clarified for this.

(working on https://github.com/openjournals/joss-reviews/issues/2842)
