### 2020-12-05 geospace-code/h5fortran#13 @bast
API documentation location could be a bit clearer/ more explicit
I was looking for an overview of the API of this library but this does not seem to be explicitly listed.

The functions can be inferred from "Examples" (which are very nice and very helpful) but it was not clear to me whether they show all available functions.

https://geospace-code.github.io/h5fortran/ lists "Procedures" and it seems it lists all public functions but this could be made clearer with a sentence or two clearly pointing the reader to where to find all interfaces documented (sorry, first time I have seen FORD so I am not familiar with this but perhaps same can be expected from an average user).

(working on https://github.com/openjournals/joss-reviews/issues/2842)
