### 2020-12-05 geospace-code/h5fortran#14 @bast
Documentation: FetchContent example in install instructions could be more complete
I was first following the FetchContent example at https://github.com/geospace-code/h5fortran/blob/master/Install.md.

This one works but does not show how to link the target against this library. I tried few things "in the blind". Eventually it worked, but I wondered why the documentation does not show a full minimal `CMakeLists.txt` example, including a `target_link_libraries`.

Later I found a nice and working full example in https://github.com/geospace-code/h5fortran/tree/master/Examples.

This means that it could be helpful for the user/reader to either be referenced to "Examples" at that point to see how to link own code within `CMakeLists.txt`, or to expand the example a bit to be self-contained.

(working on https://github.com/openjournals/joss-reviews/issues/2842)
