### 2020-12-06 geospace-code/h5fortran#10 @milancurcic
:+1: As a step further, I'd suggest allowing the user to disable sending the report, perhaps with `-DCMAKE_NO_CDASH_REPORT=1` or similar.
