### 2020-12-06 geospace-code/h5fortran#16 @milancurcic
Make Community/Contributing guidelines prominent
Currently they're hidden under .github. I almost thought there weren't any until I found them when writing a new issue. It'd be great to have them prominently available in the top-level directory and linked from the README.

openjournals/joss-reviews#2842
