### 2020-12-06 geospace-code/h5fortran#17 @bast
Documentation: statement of need and target audience
This concerns JOSS review checklist "Do the authors clearly state what problems the software is designed to solve and who the target audience is?".

I like how the README.md starts with a usage example and a one sentence summary. But I think the documentation could have 2 sentences more there clarifying what this library solves. I think this is nicely done in the paper draft "... that abstracts away most details of a frequently-used subset of HDF5 operations".

I think 2 sentences more on when to use it and what functions are available and how this is now simpler would be helpful. Thank you!

(working on https://github.com/openjournals/joss-reviews/issues/2842)
