### 2020-12-06 geospace-code/h5fortran#18 @bast
Paper draft: "References" section title seems missing
But I admit that I am not sure at what stage this gets added. But comparing with other papers this seemed missing.

(working on https://github.com/openjournals/joss-reviews/issues/2842)
