### 2020-12-10 e9cbc16
{setup,compiler_find}.cmake template
