### 2020-12-10 f1d2b73
build_hdf5: build zlib and hdf5
