+++ scripts/build_hdf5.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+
 """
 Compile HDF5 library
 
@@ -8,23 +9,22 @@ Use the full compiler path if it's not getting the right compiler.
 * FC: Fortran compiler name or path
 * CC: C compiler name or path
 """
+
 import typing as T
-import sys
 import os
 import subprocess
 import shutil
 import argparse
 import tempfile
 from pathlib import Path
+import urllib.request
+import tarfile
 
 # ========= user parameters ======================
 BUILDDIR = "build"
 HDF5_TAG = "1.10/master"
 
-
 # ========= end of user parameters ================
-
-nice = ["nice"] if sys.platform == "linux" else []
 
 
 def cli():
@@ -55,14 +55,57 @@ def cli():
     else:
         raise ValueError(f"unknown compiler {compiler}")
 
-    hdf5(
-        {"prefix": Path(prefix).expanduser(), "workdir": Path(P.workdir).expanduser()},
-        env=env,
-    )
+    dirs = {
+        "prefix": Path(prefix).expanduser(),
+        "workdir": Path(P.workdir).expanduser(),
+    }
+
+    dirs["zlib"] = zlib(dirs, env=env)
+
+    hdf5(dirs, env=env)
 
 
-def hdf5(dirs: T.Dict[str, Path], env: T.Mapping[str, str]):
-    """ build and install HDF5
+def zlib(dirs: T.Dict[str, Path], env: T.Mapping[str, str]) -> Path:
+    zlib_name = "zlib-1.2.11"
+    zlib_ext = ".tar.gz"
+    zlib_url = "https://zlib.net/" + zlib_name + zlib_ext
+
+    name = "zlib"
+
+    install_dir = dirs["prefix"] / name
+    source_dir = dirs["workdir"] / name
+    build_dir = source_dir / BUILDDIR
+
+    zlib_archive = source_dir / (zlib_name + zlib_ext)
+
+    source_dir.mkdir(exist_ok=True)
+    if not zlib_archive.is_file():
+        urllib.request.urlretrieve(zlib_url, zlib_archive)
+
+    with tarfile.open(zlib_archive) as z:
+        z.extractall(source_dir)
+
+    cmd0 = [
+        "cmake",
+        f"-S{source_dir / zlib_name}",
+        f"-B{build_dir}",
+        f"-DCMAKE_INSTALL_PREFIX={install_dir}",
+        "-DCMAKE_BUILD_TYPE=Release",
+    ]
+
+    cmd1 = ["cmake", "--build", str(build_dir), "--parallel"]
+
+    cmd2 = ["cmake", "--install", str(build_dir)]
+
+    subprocess.check_call(cmd0, env=env)
+    subprocess.check_call(cmd1)
+    subprocess.check_call(cmd2)
+
+    return install_dir
+
+
+def hdf5(dirs: T.Dict[str, Path], env: T.Dict[str, str]):
+    """build and install HDF5
     some systems have broken libz and so have trouble extracting tar.bz2 from Python.
     To avoid this, we git clone the release instead.
     """
@@ -72,9 +115,11 @@ def hdf5(dirs: T.Dict[str, Path], env: T.Mapping[str, str]):
     install_dir = dirs["prefix"] / name
     source_dir = dirs["workdir"] / name
     build_dir = source_dir / BUILDDIR
-    git_url = "https://github.com/HDFGroup/hdf5.git"
 
-    git_download(source_dir, git_url, HDF5_TAG)
+    hdf5_url = "https://github.com/HDFGroup/hdf5.git"
[…]
