### 2020-12-10 #2842 @pdebuyl
Thank you @bast for the review!

@scivision there are a number of joss review related issues opened at your repository. With JOSS, the review can happen in a mixed way between the review thread here and issues at the software's repository. You can answer directly in those issues.

In general, I prefer that an explicit notification be sent here for all steps of the review though, as I am not subscribed to the software repository and thus receive no email notification for such events.
