### 2020-12-11 geospace-code/h5fortran#18 @pdebuyl
@scivision

The last line of the article's `.md` file must be

```
# References
```
