### 2020-12-22 29eb715
cmake: accurate cpu count
