+++ Examples/CMakeLists.txt
@@ -2,6 +2,7 @@ cmake_minimum_required(VERSION 3.15)
 
 include(${CMAKE_CURRENT_SOURCE_DIR}/../cmake/policy.cmake)
 include(${CMAKE_CURRENT_SOURCE_DIR}/../cmake/compiler_find.cmake)
+find_c_fortran()
 
 project(h5fortranExample
 LANGUAGES C CXX Fortran)
