### 2021-01-03 a97225a
modernize externalProject with JSON
Build HDF5 from zip instead of git (Faster)
