### 2021-01-08 cf2a47e
handle intel windows zlib name
