### 2021-01-08 565b24a
findhdf5: don't error when some components not found
