### 2021-01-11 d4dd55d
stop compiler after 3 errors, to avoid flooding screen
