+++ cmake/compilers.cmake
@@ -7,16 +7,16 @@ include(CheckFortranCompilerFlag)
 
 if(CMAKE_Fortran_COMPILER_ID STREQUAL Intel)
   if(WIN32)
-    add_compile_options(/arch:native)
+    add_compile_options(/arch:native /Qdiag-error-limit:3)
     string(APPEND CMAKE_Fortran_FLAGS " /stand:f18 /traceback /warn /heap-arrays")
     string(APPEND CMAKE_Fortran_FLAGS_DEBUG " /check:bounds /debug:all")
   else()
-    add_compile_options(-march=native)
+    add_compile_options(-march=native -diag-error-limit=3)
     string(APPEND CMAKE_Fortran_FLAGS " -stand f18 -traceback -warn -heap-arrays")
     string(APPEND CMAKE_Fortran_FLAGS_DEBUG " -check all -debug extended")
   endif()
 elseif(CMAKE_Fortran_COMPILER_ID STREQUAL GNU)
-  add_compile_options(-mtune=native -Wall -Wextra)
+  add_compile_options(-mtune=native -Wall -Wextra -fmax-errors=3)
   string(APPEND CMAKE_Fortran_FLAGS " -fimplicit-none")
   string(APPEND CMAKE_Fortran_FLAGS_DEBUG " -fcheck=all -Werror=array-bounds")
 
