+++ cmake/compiler_find.cmake
@@ -6,38 +6,45 @@ endif()
 
 set(CMAKE_CONFIGURATION_TYPES "Release;RelWithDebInfo;Debug" CACHE STRING "Build type selections" FORCE)
 
-# Help CMake find matching compilers, especially needed for MacOS
+# --- Help CMake find matching compilers, especially needed for MacOS
 
+set(_paths)
 if(APPLE)
+  # CMAKE_SYSTEM_NAME is not set till project()
   set(_paths /usr/local/bin /opt/homebrew/bin)
   # for Homebrew that's not on PATH (can be an issue on CI)
-else()
-  set(_paths)
 endif()
 
 
 function(find_fortran)
 
-if(NOT DEFINED ENV{FC})
-  # temporarily removed ifort, because Intel oneAPI release Dec 8, 2020 is broken for HDF5 and MPI in general.
-  find_program(FC
-    NAMES gfortran gfortran-11 gfortran-10 gfortran-9 gfortran-8 gfortran-7
-    PATHS ${_paths})
-  if(FC)
-    set(ENV{FC} ${FC})
-  endif()
+set(_fc)
+if(DEFINED FC)
+  set(_fc ${FC})
+elseif(DEFINED ENV{FC})
+  set(_fc $ENV{FC})
 endif()
 
-if(NOT DEFINED ENV{FC})
-  return()
+if(_fc)
+  get_filename_component(_dir ${_fc} DIRECTORY)
 endif()
-# ensure FC exists as a executable program
+# determine if the user is intending to use Intel oneAPI or default Gfortran
+# Need to check ifort because MKLROOT may be defined for
+# use of oneMKL with Gfortran on MacOS and Linux.
+if(DEFINED ENV{MKLROOT} OR _fc MATCHES ".*ifort")
+  find_program(FC
+    NAMES ifort
+    PATHS ${_dir})
+endif()
+
 find_program(FC
-  NAMES $ENV{FC}
-  PATHS ${_paths})
+  NAMES gfortran gfortran-12 gfortran-11 gfortran-10 gfortran-9 gfortran-8 gfortran-7
+  NAMES_PER_DIR
+  PATHS ${_dir} ${_paths})
 
 if(FC)
-  set(FC ${FC} PARENT_SCOPE)
+  set(ENV{FC} ${FC})
+  # ENV{FC} is how project() picks up our hint
 endif()
 
 endfunction(find_fortran)
@@ -45,16 +52,14 @@ endfunction(find_fortran)
 
 function(find_c)
 
-if(NOT FC AND DEFINED ENV{FC})
-  set(FC $ENV{FC})
+set(_cc)
+if(DEFINED CC)
+  set(_cc ${CC})
+elseif(DEFINED ENV{CC})
+  set(_cc $ENV{CC})
 endif()
 
-message(VERBOSE " FC hint: ${FC}")
-
-if(DEFINED ENV{CC})
-  set(_name $ENV{CC})
-  set(_dir)
-else()
+if(NOT _cc)
   # remember, Apple has "/usr/bin/gcc" which is really clang
   # the technique below is NECESSARY to work on Mac and not find the wrong GCC
   if(FC)
@@ -63,26 +68,31 @@ else()
   # use same compiler for C and Fortran, which CMake might not do itself
   if(FC MATCHES ".*ifort")
     if(WIN32)
-      set(_name icl)
+      set(_cc icl)
     else()
-      set(_name icc)
+      set(_cc icc)
     endif()
   elseif(FC MATCHES ".*gfortran")
[…]
