### 2021-01-24 8a508f2
improve Zlib build and install
