+++ cmake/zlib_setup.cmake
@@ -1,34 +1,31 @@
 # build Zlib to ensure compatibility. This is a common practice for HDF5.
 
 if(WIN32)
-  set(_zlib_name ${CMAKE_STATIC_LIBRARY_PREFIX}zlibstatic${CMAKE_STATIC_LIBRARY_SUFFIX})
+  set(ZLIB_name ${CMAKE_STATIC_LIBRARY_PREFIX}zlibstatic${CMAKE_STATIC_LIBRARY_SUFFIX})
 else()
-  set(_zlib_name ${CMAKE_STATIC_LIBRARY_PREFIX}z${CMAKE_STATIC_LIBRARY_SUFFIX})
+  set(ZLIB_name ${CMAKE_STATIC_LIBRARY_PREFIX}z${CMAKE_STATIC_LIBRARY_SUFFIX})
 endif()
 
-set(_zlib_file ${PROJECT_BINARY_DIR}/ZLIBproj-prefix/src/ZLIBproj-build/${_zlib_name})
-set(_zlib_build ${PROJECT_BINARY_DIR}/ZLIBproj-prefix/src/ZLIBproj-build)
-set(_zlib_h ${PROJECT_BINARY_DIR}/ZLIBproj-prefix/src/ZLIBproj/zlib.h)
+set(ZLIB_ROOT ${PROJECT_BINARY_DIR}/zlib)
+set(ZLIB_INCLUDE_DIR ${ZLIB_ROOT}/include)
+set(ZLIB_LIBRARY ${ZLIB_ROOT}/lib/${ZLIB_name})
 
-ExternalProject_Add(ZLIBproj
+ExternalProject_Add(ZLIB
 URL ${zlib_url}
 URL_HASH SHA1=${zlib_sha1}
-UPDATE_DISCONNECTED true
-CMAKE_ARGS -DCMAKE_BUILD_TYPE=Release
-BUILD_BYPRODUCTS ${_zlib_file}
-INSTALL_COMMAND ""
+CMAKE_ARGS -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=${ZLIB_ROOT}
+BUILD_BYPRODUCTS ${ZLIB_LIBRARY}
 )
 
-# by default, Windows does not allow symbolic links, even in user directories.
-# to be safe, let's just copy the zlib.h since it's a small file.
-ExternalProject_Add_Step(ZLIBproj post-build
-COMMAND ${CMAKE_COMMAND} -E copy ${_zlib_h} ${_zlib_build}
-DEPENDEES build
-BYPRODUCTS ${_zlib_build}/zlib.h)
+
+# --- imported target
+
+file(MAKE_DIRECTORY ${ZLIB_INCLUDE_DIR})
+# avoid race condition
 
 add_library(ZLIB::ZLIB INTERFACE IMPORTED GLOBAL)
-add_dependencies(ZLIB::ZLIB ZLIBproj)  # to avoid include directory race condition
-target_link_libraries(ZLIB::ZLIB INTERFACE ${_zlib_file})
-target_include_directories(ZLIB::ZLIB INTERFACE ${_zlib_build})
+add_dependencies(ZLIB::ZLIB ZLIB)  # to avoid include directory race condition
+target_link_libraries(ZLIB::ZLIB INTERFACE ${ZLIB_LIBRARY})
+target_include_directories(ZLIB::ZLIB INTERFACE ${ZLIB_INCLUDE_DIR})
 
[…]
