### 2021-01-24 619b2f5
improve hdf5 library build
