+++ .github/workflows/ci_windows.yml
@@ -15,7 +15,12 @@ jobs:
     steps:
     - uses: actions/checkout@v2
 
-    - run: ctest -S setup.cmake -VV
+    - run: cmake -B build -G "MinGW Makefiles"
+
+    - run: cmake --build build --parallel
+
+    - run: ctest --parallel 2 --output-on-failure
+      working-directory: build
 
   # msys2:
   # works but saving CI time
