+++ cmake/build_hdf5.cmake
@@ -28,7 +28,7 @@ include(${CMAKE_CURRENT_LIST_DIR}/zlib_setup.cmake)
 ExternalProject_Add(HDF5
 URL ${hdf5_url}
 URL_HASH SHA1=${hdf5_sha1}
-CMAKE_ARGS ${zlib_root} -DCMAKE_INSTALL_PREFIX=${HDF5_ROOT} -DHDF5_GENERATE_HEADERS:BOOL=false -DHDF5_DISABLE_COMPILER_WARNINGS:BOOL=true -DBUILD_SHARED_LIBS:BOOL=false -DCMAKE_BUILD_TYPE=Release -DHDF5_BUILD_FORTRAN:BOOL=true -DHDF5_BUILD_CPP_LIB:BOOL=false -DHDF5_BUILD_TOOLS:BOOL=false -DBUILD_TESTING:BOOL=false -DHDF5_BUILD_EXAMPLES:BOOL=false
+CMAKE_ARGS ${zlib_root} -DCMAKE_INSTALL_PREFIX:PATH=${HDF5_ROOT} -DHDF5_GENERATE_HEADERS:BOOL=false -DHDF5_DISABLE_COMPILER_WARNINGS:BOOL=true -DBUILD_SHARED_LIBS:BOOL=false -DCMAKE_BUILD_TYPE=Release -DHDF5_BUILD_FORTRAN:BOOL=true -DHDF5_BUILD_CPP_LIB:BOOL=false -DHDF5_BUILD_TOOLS:BOOL=false -DBUILD_TESTING:BOOL=false -DHDF5_BUILD_EXAMPLES:BOOL=false
 BUILD_BYPRODUCTS ${HDF5_LIBRARIES}
 DEPENDS ZLIB)
 
