+++ cmake/zlib_setup.cmake
@@ -15,7 +15,7 @@ set(ZLIB_LIBRARY ${ZLIB_ROOT}/lib/${ZLIB_name})
 ExternalProject_Add(ZLIB
 URL ${zlib_url}
 URL_HASH SHA1=${zlib_sha1}
-CMAKE_ARGS -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=${ZLIB_ROOT}
+CMAKE_ARGS -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX:PATH=${ZLIB_ROOT}
 BUILD_BYPRODUCTS ${ZLIB_LIBRARY}
 )
 
