+++ .github/workflows/ci_windows.yml
@@ -15,8 +15,7 @@ jobs:
     steps:
     - uses: actions/checkout@v2
 
-    - run: cmake -B build -G "MinGW Makefiles"
-
+    - run: cmake --preset=makewin
     - run: cmake --build build --parallel
 
     - run: ctest --parallel 2 --output-on-failure
