### 2021-02-05 8144e3e
use pre-found zlib if found when building hdf5
...to avoid
add_library cannot create imported target "ZLIB::ZLIB" because another
  target with the same name already exists.
