### 2021-02-05 795a07a
FindHDF5: check that all components present before prereqs
