### 2021-02-11 33ac85f
improve install and examples
