+++ cmake/install.cmake
@@ -34,12 +34,14 @@ install(FILES
 
 # --- CPack
 
-if(PROJECT_SOURCE_DIR STREQUAL CMAKE_SOURCE_DIR)
-  set(CPACK_GENERATOR ZIP)
-  set(CPACK_PACKAGE_VENDOR "Michael Hirsch")
-  set(CPACK_PACKAGE_CONTACT "Michael Hirsch")
-  set(CPACK_DEBIAN_PACKAGE_DEPENDS "libhdf5-dev (>=1.10)")
-  set(CPACK_RESOURCE_FILE_LICENSE "${CMAKE_CURRENT_SOURCE_DIR}/LICENSE")
-  set(CPACK_RESOURCE_FILE_README "${CMAKE_CURRENT_SOURCE_DIR}/README.md")
-  include(CPack)
-endif()
+set(CPACK_GENERATOR ZIP)
+set(CPACK_PACKAGE_VENDOR "Michael Hirsch")
+set(CPACK_PACKAGE_CONTACT "Michael Hirsch")
+set(CPACK_DEBIAN_PACKAGE_DEPENDS "libhdf5-dev (>=1.10)")
+set(CPACK_RESOURCE_FILE_LICENSE "${CMAKE_CURRENT_SOURCE_DIR}/LICENSE")
+set(CPACK_RESOURCE_FILE_README "${CMAKE_CURRENT_SOURCE_DIR}/README.md")
+
+install(FILES ${CPACK_RESOURCE_FILE_README} ${CPACK_RESOURCE_FILE_LICENSE}
+  DESTINATION share/docs/${PROJECT_NAME})
+
+include(CPack)
