+++ cmake/h5fortranConfig.cmake.in
@@ -1,5 +1,3 @@
 @PACKAGE_INIT@
 
-include(${CMAKE_CURRENT_LIST_DIR}/hdf5.cmake)
-
 check_required_components(h5fortran)
