+++ .github/workflows/ci.yml
@@ -1,6 +1,7 @@
 name: ci
 env:
-  CMAKE_GENERATOR: Ninja
+   HOMEBREW_NO_INSTALL_CLEANUP: 1
+   CMAKE_GENERATOR: Ninja
 
 on:
   push:
@@ -9,6 +10,9 @@ on:
       - "**.cmake"
       - "**.f90"
       - ".github/workflows/ci.yml"
+  release:
+    types: [published]
+
 
 jobs:
 
@@ -33,6 +37,31 @@ jobs:
     - run: ctest --parallel 2 --output-on-failure
       working-directory: Examples/build
 
+    - name: create package
+      run: cpack --config build/CPackConfig.cmake
+
+    - name: Upload artifact
+      uses: actions/upload-artifact@v1
+      with:
+        name: binary-archive
+        path: build/package
+
+    - name: get version
+      id: get_version
+      run: echo ::set-output name=VERSION::${GITHUB_REF/refs\/tags\/v/}
+
+    - name: Upload release asset
+      if: github.event.action == 'published'
+      uses: actions/upload-release-asset@v1
+      env:
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      with:
+        upload_url: ${{ github.event.release.upload_url }}
+        asset_path: build/package/h5fortran-${{ steps.get_version.outputs.VERSION }}-Linux.zip
+        asset_name:  h5fortran-${{ steps.get_version.outputs.VERSION }}-Linux.zip
+        asset_content_type: application/zip
+
+
 
   mac:
     needs: linux
@@ -52,3 +81,27 @@ jobs:
     - run: cmake --build Examples/build --parallel
     - run: ctest --parallel 2 --output-on-failure
       working-directory: Examples/build
+
+    - name: create package
+      run: cpack --config build/CPackConfig.cmake
+
+    - name: Upload artifact
+      uses: actions/upload-artifact@v1
+      with:
+        name: binary-archive
+        path: build/package
+
+    - name: get version
+      id: get_version
+      run: echo ::set-output name=VERSION::${GITHUB_REF/refs\/tags\/v/}
+
+    - name: Upload release asset
+      if: github.event.action == 'published'
+      uses: actions/upload-release-asset@v1
+      env:
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      with:
+        upload_url: ${{ github.event.release.upload_url }}
+        asset_path: build/package/h5fortran-${{ steps.get_version.outputs.VERSION }}-Darwin.zip
+        asset_name:  h5fortran-${{ steps.get_version.outputs.VERSION }}-Darwin.zip
+        asset_content_type: application/zip
