### 2021-02-14 f687333
improve install and packaging
