+++ cmake/install.cmake
@@ -1,13 +1,4 @@
-# --- package specific
-
-install(FILES ${CMAKE_CURRENT_BINARY_DIR}/include/h5fortran.mod
-  DESTINATION include)
-
 # --- BOILERPLATE: install / packaging
-
-install(TARGETS ${PROJECT_NAME}
-  EXPORT ${PROJECT_NAME}Targets
-  ARCHIVE DESTINATION lib)
 
 include(CMakePackageConfigHelpers)
 
@@ -35,12 +26,19 @@ install(FILES
 # --- CPack
 
 set(CPACK_GENERATOR ZIP)
+set(CPACK_SOURCE_GENERATOR ZIP)
 set(CPACK_PACKAGE_VENDOR "Michael Hirsch")
 set(CPACK_PACKAGE_CONTACT "Michael Hirsch")
 set(CPACK_DEBIAN_PACKAGE_DEPENDS "libhdf5-dev (>=1.10)")
 set(CPACK_RESOURCE_FILE_LICENSE "${CMAKE_CURRENT_SOURCE_DIR}/LICENSE")
 set(CPACK_RESOURCE_FILE_README "${CMAKE_CURRENT_SOURCE_DIR}/README.md")
 set(CPACK_OUTPUT_FILE_PREFIX "${CMAKE_CURRENT_BINARY_DIR}/package")
+set(CPACK_PACKAGE_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR})
+
+# not .gitignore as its regex syntax is more advanced than CMake
+file(READ ${CMAKE_CURRENT_LIST_DIR}/.cpack_ignore _cpack_ignore)
+string(REGEX REPLACE "\n" ";" _cpack_ignore ${_cpack_ignore})
+set(CPACK_SOURCE_IGNORE_FILES "${_cpack_ignore}")
 
 install(FILES ${CPACK_RESOURCE_FILE_README} ${CPACK_RESOURCE_FILE_LICENSE}
   DESTINATION share/docs/${PROJECT_NAME})
