### 2021-02-17 0f633ef
compact dataset prototype benchmark
need to benchmark on multiple OS and compilers.
Windows GCC HDF5 1.12.0 shows compact write ~ 3x faster than contiguous and chunk
