+++ src/concepts/h5compact.f90
@@ -1,0 +1,230 @@
+!! Example of compact vs. contiguous datasets, requires HDF5 >= 1.8
+!! for convenience of comparison we create two files,
+!! though distinct dataset types can exist in the same file.
+!!
+!! The file size saving isn't so much, but the advantage is in speed of access
+!! particularly when small datasets are frequently read/write
+
+module h5layout
+
+use, intrinsic :: iso_fortran_env, only : REAL64,INT64, stderr=>error_unit
+use hdf5
+use h5lt, only : h5ltread_dataset_float_f
+
+implicit none (type, external)
+
+private
+public :: print_layout, write_layout, sysclock2ms, unlink, bench_write, bench_read
+
+contains
+
+subroutine print_layout(filename, ds_name)
+!! prints dataset (variable) layout in a file
+character(*), intent(in) :: filename, ds_name
+
+integer :: ierr, layout
+integer(hid_t) :: f_id, ds_id, prp_id
+INTEGER(HSIZE_T) :: fsize
+
+CALL h5fopen_f(filename, H5F_ACC_RDONLY_F, f_id, ierr)
+
+!> get file size.
+CALL h5fget_filesize_f(f_id, fsize, ierr)
+print *, filename," size in bytes: ", fsize
+
+!> get dataset storage layout
+CALL h5dopen_f(f_id, ds_name, ds_id, ierr)
+if (ierr /= 0 ) error stop "problem opening dataset"
+CALL h5dget_create_plist_f(ds_id, prp_id, ierr)
+CALL h5pget_layout_f(prp_id, layout, ierr)
+
+!> not truly constants, so select case doesn't work
+if (layout == H5D_COMPACT_F) then
+  print *, ds_name, " layout: COMPACT"
+elseif(layout == H5D_CONTIGUOUS_F) then
+  print *, ds_name, " layout: CONTIGUOUS"
+elseif(layout == H5D_CHUNKED_F) then
+  print *, ds_name, " layout: CHUNKED"
+else
+  error stop "could not determine layout"
+end if
+
+call h5pclose_f(prp_id, ierr)
+call h5dclose_f(ds_id, ierr)
+call h5fclose_f(f_id, ierr)
+CALL h5close_f(ierr)
+
+end subroutine print_layout
+
+
+subroutine write_layout(file_id, ds_name, layout, dat)
+!! write real data with HDF5 layout type specified
+
+integer(hid_t), intent(in) :: file_id
+character(*), intent(in) :: ds_name
+integer, intent(in) :: layout
+real, intent(in) :: dat(:)
+
+integer :: ierr
+integer(hid_t) :: ds_id, prp_id, space_id
+integer(HSIZE_T) :: dims(rank(dat))
+
+dims = shape(dat)
+
+!> create dataspace
+call h5screate_simple_f(size(dims), dims, space_id, ierr)
+if (ierr /= 0) error stop "problem creating dataspace"
+
+!> create dataset property list
+call h5pcreate_f(H5P_DATASET_CREATE_F, prp_id, ierr)
+if (ierr /= 0) error stop "problem creating property list"
+
+!> config for dataset (must be before h5pset_chunk)
+call h5pset_layout_f(prp_id, layout, ierr)
+if (ierr /= 0) error stop "problem creating layout"
+
+if(layout == H5D_CHUNKED_F) then
+   !! normally chunk size set more effectively for actual large datasets.
+   call h5pset_chunk_f(prp_id, size(dims), int(shape(dat), hsize_t), ierr)
[…]
