+++ src/CMakeLists.txt
@@ -1,5 +1,6 @@
-# add_subdirectory(concepts)
-
+if(concepts)
+add_subdirectory(concepts)
+endif(concepts)
 
 if(WIN32)
   set(is_windows .true.)
