### 2021-02-19 353c394
add C-Fortran ABI check
ci: use presets
