+++ cmake/abi_check/CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.14)
+project(abi_check LANGUAGES C Fortran)
+
+add_library(addone OBJECT addone.c)
+
+add_executable(main main.f90 $<TARGET_OBJECTS:addone>)
