+++ CMakeLists.txt
@@ -1,12 +1,8 @@
 cmake_minimum_required(VERSION 3.13)
 
-include(cmake/policy.cmake)
-include(cmake/compiler_find.cmake)
-find_c_fortran()
-
 project(h5fortran
   LANGUAGES C Fortran
-  VERSION 3.4.10
+  VERSION 3.4.11
   DESCRIPTION "thin, light object-oriented HDF5 Fortran interface"
   HOMEPAGE_URL https://github.com/geospace-code/h5fortran)
 
@@ -17,6 +13,7 @@ endif()
 
 list(APPEND CMAKE_MODULE_PATH ${CMAKE_CURRENT_SOURCE_DIR}/cmake/Modules)
 
+include(cmake/policy.cmake)
 include(cmake/options.cmake)
 include(cmake/compilers.cmake)
 include(cmake/libraries.cmake)
