### 2021-02-21 90cae2e
eliminate workaround and extra options
