+++ cmake/options.cmake
@@ -7,6 +7,5 @@ option(concepts "conceptual testing, for devs only" off)
 set(CMAKE_EXPORT_COMPILE_COMMANDS on)
 
 if(NOT dev)
-  set(FETCHCONTENT_UPDATES_DISCONNECTED_ZLIB true)
-  set(FETCHCONTENT_UPDATES_DISCONNECTED_HDF5 true)
+  set(EP_UPDATE_DISCONNECTED true)
 endif()
