+++ cmake/h5fortranConfig.cmake.in
@@ -1,3 +1,3 @@
 @PACKAGE_INIT@
 
-check_required_components(h5fortran)
+check_required_components(${PROJECT_NAME})
