+++ cmake/zlib_setup.cmake
@@ -15,6 +15,9 @@ set(ZLIB_LIBRARY ${ZLIB_ROOT}/lib/${ZLIB_name})
 ExternalProject_Add(ZLIB
 URL ${zlib_url}
 URL_HASH SHA1=${zlib_sha1}
+TLS_VERIFY ON
+CONFIGURE_HANDLED_BY_BUILD ON
+INACTIVITY_TIMEOUT 30
 CMAKE_ARGS -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX:PATH=${ZLIB_ROOT}
 BUILD_BYPRODUCTS ${ZLIB_LIBRARY}
 )

[…]
