### 2021-02-21 01f00ce
add COMPACT write
