+++ .github/workflows/ci.yml
@@ -62,6 +62,27 @@ jobs:
         asset_content_type: application/zip
 
 
+  linuxCmake313:
+    runs-on: ubuntu-20.04
+    steps:
+    - uses: actions/checkout@v2
+
+    - name: install hdf5
+      run: |
+        sudo apt update -yq
+        sudo apt install -yq --no-install-recommends ninja-build gfortran libhdf5-dev
+
+    - name: get CMake 3.13
+      run: |
+        curl -L -O https://github.com/Kitware/CMake/releases/download/v3.13.5/cmake-3.13.5-Linux-x86_64.tar.gz
+        tar xf cmake-3.13.5-Linux-x86_64.tar.gz
+        echo "${GITHUB_WORKSPACE}/cmake-3.13.5-Linux-x86_64/bin" >> $GITHUB_PATH
+
+    - run: cmake -B build
+    - run: cmake --build build --parallel
+    - run: ctest --parallel 2 --output-on-failure
+      working-directory: build
+
 
   mac:
     needs: linux
