+++ cmake/abi_check/CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.14)
+cmake_minimum_required(VERSION 3.13)
 project(abi_check LANGUAGES C Fortran)
 
 add_library(addone OBJECT addone.c)
