### 2021-02-22 42a1eb6
read/ write/ src directories for orderliness. test files not at build root
