+++ src/write/CMakeLists.txt
@@ -1,0 +1,10 @@
+file(READ writer_template.in.f90 writer_template)
+configure_file(writer.in.f90 writer.f90 @ONLY)
+
+file(READ writer_lt_template.in.f90 writer_lt_template)
+configure_file(writer_lt.in.f90 writer_lt.f90 @ONLY)
+
+target_sources(h5fortran PRIVATE
+  write.f90
+  ${CMAKE_CURRENT_BINARY_DIR}/writer.f90
+  ${CMAKE_CURRENT_BINARY_DIR}/writer_lt.f90)
