### 2021-02-22 f1746b5
check dataset type/kind before reading
