### 2021-02-22 29ebb29
endianness check thought not needed
