+++ src/read/read.f90
@@ -19,7 +19,8 @@ character(*), intent(in) :: dname, filename
 
 integer(hid_t) :: dtype_id, native_dtype_id
 integer :: class
-integer :: ierr, order, machine_order
+integer :: ierr
+! integer :: order, machine_order
 integer(size_t) :: size_bytes
 
 !> get the dataset variable type
@@ -32,11 +33,12 @@ if(ierr/=0) error stop 'h5fortran:reader: get internal dtype ' // dname // ' fro
 call h5tget_native_type_f(dtype_id, H5T_DIR_ASCEND_F, native_dtype_id, ierr)
 if(ierr/=0) error stop 'h5fortran:reader: get native dtype id ' // dname // ' from ' // filename
 
-call h5tget_order_f(native_dtype_id, order, ierr)
-if(ierr/=0) error stop 'h5fortran:reader: get endianness ' // dname // ' from ' // filename
-!> check dataset endianness matches machine (in future, could swap endianness if needed)
-call h5tget_order_f(H5T_NATIVE_INTEGER, machine_order, ierr)
-if(order /= machine_order) error stop 'h5fortran:reader: endianness does not match machine native ' // dname // ' from ' // filename
+!> we think endianness is handled by HDF5 ... ?
+! call h5tget_order_f(native_dtype_id, order, ierr)
+! if(ierr/=0) error stop 'h5fortran:reader: get endianness ' // dname // ' from ' // filename
+! !> check dataset endianness matches machine (in future, could swap endianness if needed)
+! call h5tget_order_f(H5T_NATIVE_INTEGER, machine_order, ierr)
+! if(order /= machine_order) error stop 'h5fortran:reader: endianness does not match machine native ' // dname // ' from ' // filename
 
 !> compose datatype inferred
 call h5tget_class_f(native_dtype_id, class, ierr)
