+++ CMakeLists.txt
@@ -60,6 +60,7 @@ if(${PROJECT_NAME}_BUILD_TESTING)
   add_subdirectory(src/tests)
 endif()
 
+include(cmake/pkgconf.cmake)
 include(cmake/install.cmake)
 
 include(cmake/summary.cmake)
