+++ .github/workflows/ci.yml
@@ -57,8 +57,8 @@ jobs:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
       with:
         upload_url: ${{ github.event.release.upload_url }}
-        asset_path: build/package/h5fortran-${{ steps.get_version.outputs.VERSION }}-Linux.zip
-        asset_name:  h5fortran-${{ steps.get_version.outputs.VERSION }}-Linux.zip
+        asset_path: build/package/h5fortran-${{ steps.get_version.outputs.VERSION }}-Linux.tar.gz
+        asset_name:  h5fortran-${{ steps.get_version.outputs.VERSION }}-Linux.tar.gz
         asset_content_type: application/zip
 
 
@@ -126,6 +126,6 @@ jobs:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
       with:
         upload_url: ${{ github.event.release.upload_url }}
-        asset_path: build/package/h5fortran-${{ steps.get_version.outputs.VERSION }}-Darwin.zip
-        asset_name:  h5fortran-${{ steps.get_version.outputs.VERSION }}-Darwin.zip
+        asset_path: build/package/h5fortran-${{ steps.get_version.outputs.VERSION }}-Darwin.tar.gz
+        asset_name:  h5fortran-${{ steps.get_version.outputs.VERSION }}-Darwin.tar.gz
         asset_content_type: application/zip
