+++ CMakeLists.txt
@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 3.13...3.20)
 
 project(h5fortran
   LANGUAGES C Fortran
-  VERSION 3.5.2
+  VERSION 3.5.3
   DESCRIPTION "thin, light object-oriented HDF5 Fortran interface"
   HOMEPAGE_URL https://github.com/geospace-code/h5fortran)
 
