+++ cmake/compilers.cmake
@@ -27,8 +27,8 @@ if(CMAKE_Fortran_COMPILER_ID STREQUAL Intel)
     string(APPEND CMAKE_Fortran_FLAGS_DEBUG " -check all -debug extended")
   endif()
 elseif(CMAKE_Fortran_COMPILER_ID STREQUAL GNU)
-  string(APPEND CMAKE_Fortran_FLAGS " -mtune=native -Wall -Wextra -fimplicit-none")
-  string(APPEND CMAKE_Fortran_FLAGS_DEBUG " -fcheck=all -Werror=array-bounds")
+  string(APPEND CMAKE_Fortran_FLAGS " -mtune=native -Wall -fimplicit-none")
+  string(APPEND CMAKE_Fortran_FLAGS_DEBUG " -Wextra -fcheck=all -Werror=array-bounds")
 
   # check_fortran_compiler_flag(-std=f2018 f2018flag)
   # if(f2018flag)
