### 2021-03-03 35acf89
allow user to specify HDF5_ROOT and ZLIB_ROOT when building
