+++ setup.cmake
@@ -1,30 +1,70 @@
-# run by:
-# ctest -S setup.cmake
+cmake_minimum_required(VERSION 3.14...3.20)
 
-# --- Project-specific -Doptions
-# these will be used if the project isn't already configured.
-set(_opts)
+set(CTEST_PROJECT_NAME "h5fortran")
+set(CTEST_NIGHTLY_START_TIME "01:00:00 UTC")
+set(CTEST_MODEL "Experimental")
+set(CTEST_SUBMIT_URL "https://my.cdash.org/submit.php?project=${CTEST_PROJECT_NAME}")
+
+set(CTEST_LABELS_FOR_SUBPROJECTS "unit;shaky")
 
 # --- boilerplate follows
-message(STATUS "CMake ${CMAKE_VERSION}")
-if(CMAKE_VERSION VERSION_LESS 3.15)
-  message(FATAL_ERROR "Please update CMake >= 3.15.
-    Try 'pip install -U cmake' or https://cmake.org/download/")
+set(CTEST_TEST_TIMEOUT 10)
+set(CTEST_OUTPUT_ON_FAILURE true)
+
+set(CTEST_SOURCE_DIRECTORY ${CTEST_SCRIPT_DIRECTORY})
+if(NOT DEFINED CTEST_BINARY_DIRECTORY)
+  set(CTEST_BINARY_DIRECTORY ${CTEST_SOURCE_DIRECTORY}/build)
+endif()
+
+if(NOT DEFINED CTEST_BUILD_CONFIGURATION)
+  set(CTEST_BUILD_CONFIGURATION "Release")
+endif()
+
+if(NOT DEFINED CTEST_SITE)
+  if(DEFINED ENV{CTEST_SITE})
+    set(CTEST_SITE $ENV{CTEST_SITE})
+  else()
+    cmake_host_system_information(RESULT sys_name QUERY OS_NAME OS_RELEASE OS_VERSION)
+    string(REPLACE ";" " " sys_name ${sys_name})
+    set(CTEST_SITE ${sys_name})
+  endif()
+endif()
+
+if(NOT DEFINED CTEST_BUILD_NAME)
+  if(DEFINED ENV{CTEST_BUILD_NAME})
+    set(CTEST_BUILD_NAME $ENV{CTEST_BUILD_NAME})
+  else()
+    find_program(GIT_EXECUTABLE NAMES git REQUIRED)
+    execute_process(COMMAND ${GIT_EXECUTABLE} describe --tags
+      WORKING_DIRECTORY ${CTEST_SOURCE_DIRECTORY}
+      OUTPUT_VARIABLE git_rev OUTPUT_STRIP_TRAILING_WHITESPACE
+      RESULT_VARIABLE _err)
+    if(NOT _err EQUAL 0)
+      # old Git
+      execute_process(COMMAND ${GIT_EXECUTABLE} rev-parse --short HEAD
+        WORKING_DIRECTORY ${CTEST_SOURCE_DIRECTORY}
+        OUTPUT_VARIABLE git_rev OUTPUT_STRIP_TRAILING_WHITESPACE
+        RESULT_VARIABLE _err)
+    endif()
+    if(_err EQUAL 0)
+      set(CTEST_BUILD_NAME ${git_rev})
+    endif()
+  endif()
 endif()
 
 # CTEST_CMAKE_GENERATOR must always be defined
 if(NOT DEFINED CTEST_CMAKE_GENERATOR AND CMAKE_VERSION VERSION_GREATER_EQUAL 3.17)
-  find_program(_gen NAMES ninja ninja-build samu)
-  if(_gen)
-    execute_process(COMMAND ${_gen} --version
[…]
