### 2021-03-12 75193ab
build_hdf5.py: allow running from same directory
