### 2021-03-12 5bd6657
build_hdf5.py: command line option to not download release
this is relevant when Github release downloads aren't available,
like today!
