### 2021-03-12 78173f5
findHDF5: link test needs libm on Unix
