### 2021-03-12 12a22e2
ctest setup.cmake: use ENV{CMAKE_GENERATOR} if present
