+++ setup.cmake
@@ -52,30 +52,14 @@ if(NOT DEFINED CTEST_BUILD_NAME)
   endif()
 endif()
 
-# for subproject labels
-set(CTEST_USE_LAUNCHERS 1)
-set(ENV{CTEST_USE_LAUNCHERS_DEFAULT} 1)
 
-# CTEST_CMAKE_GENERATOR must always be defined
-if(NOT DEFINED CTEST_CMAKE_GENERATOR AND CMAKE_VERSION VERSION_GREATER_EQUAL 3.17)
-  find_program(ninja NAMES ninja ninja-build samu)
-  if(ninja)
-    execute_process(COMMAND ${ninja} --version
-      OUTPUT_VARIABLE ninja_version
-      OUTPUT_STRIP_TRAILING_WHITESPACE
-      RESULT_VARIABLE err
-      TIMEOUT 10)
-    if(err EQUAL 0 AND ninja_version VERSION_GREATER_EQUAL 1.10)
-      set(CTEST_CMAKE_GENERATOR Ninja)
-    endif()
-  endif(ninja)
-endif()
+set(CTEST_USE_LAUNCHERS 1)
+
 if(NOT DEFINED CTEST_CMAKE_GENERATOR)
-  set(CTEST_BUILD_FLAGS -j)  # not --parallel as this goes to generator directly
-  if(WIN32)
-    set(CTEST_CMAKE_GENERATOR "MinGW Makefiles")
+  if(DEFINED ENV{CMAKE_GENERATOR})
+    set(CTEST_CMAKE_GENERATOR $ENV{CMAKE_GENERATOR})
   else()
-    set(CTEST_CMAKE_GENERATOR "Unix Makefiles")
+    include(cmake/find_generator.cmake)
   endif()
 endif()
 
