+++ src/interface.f90
@@ -95,9 +95,9 @@ end interface h5read_attr
 !> Submodules
 
 interface !< pathlib.f90
-module logical function unlink(filename)
+module logical function std_unlink(filename)
 character(*), intent(in) :: filename
-end function unlink
+end function std_unlink
 
 module logical function is_absolute_path(path)
 character(*), intent(in) :: path
@@ -629,7 +629,7 @@ endif
 self%lid = 0
 
 if(self%is_scratch) then
-  if (unlink(self%filename)) write(stderr,*) 'WARNING: could not delete scratch file: ' // self%filename
+  if (std_unlink(self%filename)) write(stderr,*) 'WARNING: could not delete scratch file: ' // self%filename
 endif
 
 self%is_open = .false.
