+++ cmake/compilers.cmake
@@ -28,17 +28,9 @@ if(CMAKE_Fortran_COMPILER_ID STREQUAL Intel)
   endif()
 elseif(CMAKE_Fortran_COMPILER_ID STREQUAL GNU)
   string(APPEND CMAKE_Fortran_FLAGS " -mtune=native -Wall -fimplicit-none")
+  string(APPEND CMAKE_Fortran_FLAGS_RELEASE " -Wno-maybe-uninitialized")
+  string(APPEND CMAKE_Fortran_FLAGS_RELWITHDEBINFO " -Wno-maybe-uninitialized")
   string(APPEND CMAKE_Fortran_FLAGS_DEBUG " -Wextra -fcheck=all -Werror=array-bounds")
-
-  # check_fortran_compiler_flag(-std=f2018 f2018flag)
-  # if(f2018flag)
-  #   string(APPEND CMAKE_Fortran_FLAGS " -std=f2018")
-  # endif()
-
-  if(CMAKE_Fortran_COMPILER_VERSION VERSION_EQUAL 9.3.0)
-    # makes a lot of spurious warnings on allocatable scalar character
-    string(APPEND CMAKE_Fortran_FLAGS " -Wno-maybe-uninitialized")
-  endif()
 elseif(CMAKE_Fortran_COMPILER_ID STREQUAL NAG)
   string(APPEND CMAKE_Fortran_FLAGS " -f2018 -u -C=all")
 endif()
