### 2021-03-12 1821dfb
setup.cmake: handle CI nightly/continuous cases
