### 2021-03-12 46a7de5
setup.cmake: ensure build_type is enacted
