+++ .github/workflows/ci_windows.yml
@@ -1,6 +1,7 @@
 name: ci_windows
 env:
   CMAKE_GENERATOR: Ninja
+  CTEST_SITE: GitHubActions
 
 on:
   push:
