### 2021-03-12 4009f59
ctest: warn instead of fail if not git porcelain
