+++ setup.cmake
@@ -121,13 +121,13 @@ if(CTEST_MODEL STREQUAL Nightly OR CTEST_MODEL STREQUAL Continuous)
     message(FATAL_ERROR "CTest could not check Git porcelain status")
   endif()
   if(_ret)
-    message(FATAL_ERROR "CTest would have erased the non-Git Push'd changes, aborting instead of erasing.")
-  endif()
-
-  ctest_update(RETURN_VALUE _ret)
-  if(_ret EQUAL 0)
-    message(STATUS "No Git-updated files, so no need to test. CTest stopping.")
-    return()
+    message(WARNING "CTest would have erased the non-Git Push'd changes, NOT updating.")
+  else()
+    ctest_update(RETURN_VALUE _ret)
+    if(_ret EQUAL 0)
+      message(STATUS "No Git-updated files, so no need to test. CTest stopping.")
+      return()
+    endif()
   endif()
 endif()
 
