+++ .github/workflows/ci.yml
@@ -25,7 +25,7 @@ jobs:
         sudo apt update -yq
         sudo apt install -yq --no-install-recommends ninja-build gfortran libhdf5-dev
 
-    - run: ctest -S setup.cmake -V
+    - run: ctest -S ci.cmake -V
 
     - run: cmake -DCMAKE_INSTALL_PREFIX=~/hdf5
     - run: cmake --install build
@@ -98,7 +98,7 @@ jobs:
 
     - run: brew install hdf5 ninja
 
-    - run: ctest -S setup.cmake -V
+    - run: ctest -S ci.cmake -V
 
     # - run: cmake -S Examples -B Examples/build
     # - run: cmake --build Examples/build --parallel
