+++ .github/workflows/ci_windows.yml
@@ -48,7 +48,7 @@ jobs:
 
     - uses: actions/checkout@v2
 
-    - run: ctest -S setup.cmake -V
+    - run: ctest -S ci.cmake -V
 
     # - run: cmake -S Examples -B Examples/build
     # - run: cmake --build Examples/build --parallel
