+++ ci.cmake
@@ -107,6 +107,7 @@ cmake_cpu_count()
 
 set(CTEST_NOTES_FILES "${CTEST_SCRIPT_DIRECTORY}/${CTEST_SCRIPT_NAME}")
 set(CTEST_SUBMIT_RETRY_COUNT 3)
+set(CTEST_UPDATE_COMMAND git)  # avoids auto-detect version control failures on some systemes
 
 ctest_start(${CTEST_MODEL})
 
