### 2021-03-12 b2c50bb
CI nightly mode as CI doesn't see updates from fresh clone
