+++ ci.cmake
@@ -20,7 +20,7 @@ if(NOT CTEST_MODEL)
   if(DEFINED ENV{CI})
     set(CI $ENV{CI})
     if(CI)
-      set(CTEST_MODEL "Continuous")
+      set(CTEST_MODEL "Nightly")
     endif()
   endif()
 endif()
@@ -125,8 +125,8 @@ if(CTEST_MODEL STREQUAL Nightly OR CTEST_MODEL STREQUAL Continuous)
     message(WARNING "CTest would have erased the non-Git Push'd changes, NOT updating.")
   else()
     ctest_update(RETURN_VALUE _ret)
-    if(_ret EQUAL 0)
-      message(STATUS "No Git-updated files, so no need to test. CTest stopping.")
+    if(_ret EQUAL 0 AND CTEST_MODEL STREQUAL Continuous)
+      message(STATUS "No Git-updated files, so no need to test in CTest Model ${CTEST_MODEL}. CTest stopping.")
       return()
     endif()
   endif()
