+++ .github/workflows/ci_windows.yml
@@ -29,7 +29,7 @@ jobs:
   #     working-directory: build
 
   msys2:
-    timeout-minutes: 15
+    timeout-minutes: 20
     runs-on: windows-latest
 
     defaults:
@@ -48,12 +48,18 @@ jobs:
 
     - uses: actions/checkout@v2
 
-    - run: ctest -S ci.cmake -V
+    - name: CDash
+      run: ctest -S ci.cmake -V
 
-    # - run: cmake -S Examples -B Examples/build
-    # - run: cmake --build Examples/build --parallel
-    # - run: ctest --parallel 2 --output-on-failure
-    #   working-directory: Examples/build
+    - name: Install package
+      run: |
+        cmake -B build -DCMAKE_INSTALL_PREFIX=~/h5fortran
+        cmake --install build
+
+    - run: cmake -S Examples -B Examples/build -Dh5fortran_ROOT=~/h5fortran
+    - run: cmake --build Examples/build --parallel
+    - run: ctest --parallel 2 --output-on-failure
+      working-directory: Examples/build
 
     - name: create package
       if: github.event.action == 'published'
