### 2021-03-14 5d81b65
ctest need ctest_update_command to avoid final return code 255
