### 2021-03-15 550ea07
ci.cmake one file
version from file, install Find*.cmake
