### 2021-03-15 f6bf95f
install: find HDF5 automatically
packages using h5fortran install no longer need to find HDF5 themselves
