### 2021-03-16 079793f
ci.cmake template: arg passing
