### 2021-03-22 geospace-code/h5fortran#10 @scivision
Thanks, I removed that script from the user documentation and instead only use it for CI.
