### 2021-03-22 b05c8f3
update install docs
