### 2021-03-22 geospace-code/h5fortran#11 @scivision
Yes thank you. I removed the inconsistencies and deduplicated text: https://github.com/geospace-code/h5fortran/blob/main/Install.md#use-h5fortran-from-your-project

The new h5fortran packaging method automatically seeks HDF5 library, the user doesn't need to manually specify that in their own project's CMakeLists.txt
