### 2021-03-22 geospace-code/h5fortran#17 @scivision
thanks this is done in https://github.com/geospace-code/h5fortran/commit/593f63693a30b90cb0d681f074a9b2cc1eb80742
