### 2021-03-22 geospace-code/h5fortran#16 @scivision
Thanks I made a link. I use organization-wide templates due to the large number of repositories I have. https://github.com/geospace-code/h5fortran/commit/a8b556260705bad89386bae20e9504c82f42f26d
