### 2021-03-22 geospace-code/h5fortran#12 @scivision
This was fixed in the meantime, thanks!
