### 2021-03-22 b373a18
make install.md a self-contained example
