### 2021-03-22 geospace-code/h5fortran#14 @scivision
Thanks I made this in https://github.com/geospace-code/h5fortran/blob/main/Install.md#use-h5fortran-from-your-project
