### 2021-03-22 geospace-code/h5fortran#15 @scivision
Yes I added to readme.md, that if one isn't using CMake for their own project, they can use the HDF5 compiler wrapper like:

```sh
h5fc -I~/h5fortran/include myprogram.f90 ~/h5fortran/lib/libh5fortran.a
```
