+++ src/test_array.f90
@@ -1,25 +1,38 @@
-module test_array
+program array_test
 
 use, intrinsic:: ieee_arithmetic, only: ieee_value, ieee_quiet_nan, ieee_is_nan
 use, intrinsic :: iso_fortran_env, only: real32, real64, int32, stderr=>error_unit
-use h5fortran, only : hdf5_file, hsize_t
+use h5fortran, only : hdf5_file, HSIZE_T, H5T_NATIVE_INTEGER
 
-implicit none
+implicit none (type, external)
 
 real(real32) :: nan
 
+call test_basic_array('test_array.h5')
+print *,'PASSED: array write'
+call test_read_slice('test_array.h5')
+print *, 'PASSED: slice read'
+call test_write_slice('test_array.h5')
+print *, 'PASSED: slice write'
+
+call test_readwrite_array('test_group_array.f90', ng=69, nn=100, pn=5)
+print *,'PASSED: array write / read'
+
+
 contains
 
-subroutine test_write_array(path)
+subroutine test_basic_array(filename)
+
+character(*), intent(in) :: filename
 !! tests that compression doesn't fail for very small datasets, where it really shouldn't be used (makes file bigger)
-type(hdf5_file) :: h5f
-character(*), intent(in) :: path
+type(hdf5_file) :: h
+integer(HSIZE_T), allocatable :: dims(:)
 
 integer(int32), dimension(4) :: i1, i1t
 integer(int32), dimension(4,4) :: i2, i2t
 real(real32), allocatable :: rr2(:,:)
-real(real32)  ::  nant, r1(4), r2(4,4)
-integer :: i, ierr
+real(real32)  ::  nant, r1(4), r2(4,4), B(6,6)
+integer :: i
 integer(int32) :: i2_8(8,8)
 
 nan = ieee_value(1.0, ieee_quiet_nan)
@@ -36,121 +49,163 @@ enddo
 r1 = i1
 r2 = i2
 
-!! write test data
-call h5f%initialize(path//'/test.h5', status='old',action='rw',comp_lvl=1, verbose=.False.)
+call h%initialize(filename, status='replace', comp_lvl=1, verbose=.False.)
 
-call h5f%write('/int32-1d', i1)
-call h5f%write('/test/group2/int32-2d', i2)
-call h5f%write('/test/real2', r2)
-call h5f%write('/nan', nan)
+call h%write('/int32-1d', i1)
+call h%write('/test/group2/int32-2d', i2)
+call h%write('/real32-2d', r2)
+call h%write('/nan', nan)
 
-!> test writing wrong size
-call h5f%write('/int32-1d', [-1], ierr)
-if(ierr==0) error stop 'test_write_array: did not error for write array shape mismatch'
+call h%finalize()
 
-!> test writing wrong rank
-call h5f%write('/int32-1d', i2, ierr)
-if(ierr==0) error stop 'test_write_array: did not error for write array rank mismatch'
+!! read
+call h%initialize(filename, status='old', action='r', verbose=.false.)
 
-call h5f%finalize()
-
-!! Read tests
-call h5f%initialize(path//'/test.h5', status='old',action='r', verbose=.false.)
 !> int32
-
-call h5f%read('/int32-1d', i1t)
+call h%read('/int32-1d', i1t)
 if (.not.all(i1==i1t)) error stop 'read 1-d int32: does not match write'
 
-print *, 'test_write_array: read slice 1d, stride=1'
-i1t = 0
-call h5f%read('/int32-1d', i1t(:2), istart=[2], iend=[3], stride=[1])
[…]
