+++ src/test_error.f90
@@ -3,138 +3,66 @@ program test_error
 use, intrinsic:: iso_fortran_env, only: int32, real32, real64, stderr=>error_unit
 use h5fortran, only: hdf5_file
 
-implicit none
+implicit none (type, external)
 
-type(hdf5_file) :: h5f
-
-character(:), allocatable :: path
-character(256) :: argv
-integer :: i,l
-
-call get_command_argument(1, argv, length=l, status=i)
-if (i /= 0 .or. l == 0) then
-  write(stderr,*) 'please specify test directory e.g. /tmp'
-  error stop 77
-endif
-path = trim(argv)
-print *, 'test path: ', path
-
-call test_nonexist_old_file(path)
+call test_nonexist_old_file()
 print *, 'OK: non-existing old file'
-call test_nonexist_unknown_file(path)
+call test_nonexist_unknown_file()
 print *, 'OK: non-existing unknown file'
-call test_nonhdf5_file(path)
+call test_nonhdf5_file()
 print *, 'OK: non-HDF5 file'
-call test_nonexist_variable(path)
-print *, 'OK: non-existing variable'
-call test_wrong_type(path)
+call test_wrong_type()
 print *, "OK: wrong type read"
-call test_unknown_write(path)
-print *, 'OK: unknown write'
-call test_unknown_read(path)
-print *, 'OK: unknown read'
 
 contains
 
 
-subroutine test_nonexist_old_file(path)
-character(*), intent(in) :: path
+subroutine test_nonexist_old_file()
 integer :: ierr
+type(hdf5_file) :: h
 
-call h5f%initialize(path//'/not-exist', ierr, status='old', action='read', verbose=.false.)
+call h%initialize('not-exist.h5', ierr, status='old', action='read', verbose=.false.)
 if (ierr==0) error stop 'should have had ierr/=0 on non-existing old file'
 end subroutine test_nonexist_old_file
 
 
-subroutine test_nonexist_unknown_file(path)
-character(*), intent(in) :: path
+subroutine test_nonexist_unknown_file()
 integer :: ierr
+type(hdf5_file) :: h
 
-call h5f%initialize(path//'/not-exist', ierr, status='unknown', action='read', verbose=.false.)
+call h%initialize('not-exist.h5', ierr, status='unknown', action='read', verbose=.false.)
 if (ierr==0) error stop 'should have had ierr/=0 on non-existing unknown read file'
 end subroutine test_nonexist_unknown_file
 
 
-subroutine test_nonhdf5_file(path)
-character(*), intent(in) :: path
+subroutine test_nonhdf5_file()
 integer :: u,ierr
-character(:), allocatable :: filename
-
-filename = path // '/junk.dat'
+type(hdf5_file) :: h
+character(*), parameter :: filename = 'bad.h5'
 
 ! create or replace zero-length file, could be any size, just not a valid HDF5 file
-open(newunit=u, file=filename, status='replace', iostat=ierr, action='write')
+open(newunit=u, file=filename, status='replace', action='write')
 close(u)
 
-call h5f%initialize(filename, ierr, status='old', action='read')
+call h%initialize(filename, ierr, status='old', action='read')
 if (ierr==0) error stop 'should have had ierr/=0 on invalid HDF5 file'
 end subroutine test_nonhdf5_file
 
 
-subroutine test_nonexist_variable(path)
-character(*), intent(in) :: path
-integer :: u,ierr
-character(:), allocatable :: filename
-
-filename = path // '/junk.h5'
-call h5f%initialize(filename, status='replace', action='readwrite', verbose=.false.)
-call h5f%read('/not-exist', u, ierr)
[…]
