+++ API.md
@@ -6,6 +6,7 @@ All examples assume:
 
 ```fortran
 use h5fortran, only: hdf5_file
+use hdf5, only: HSIZE_T, HID_T
 
 type(hdf5_file) :: h
 ```
@@ -46,10 +47,145 @@ logical, intent(in), optional :: close_hdf5_interface
 
 To avoid memory leaks or corrupted files, always "finalize" all hDF5 files before STOPping the Fortran program.
 
-
 ```fortran
 call h%flush(ierr)
 !! request operating system flush data to disk.
 !! The operating system can do this when it desires, which might be a while.
+integer, intent(out), optional :: ierr
+```
+
+## Disk variable (dataset) inquiry
+
+To allocate variables before reading data, inquire about dataset characteristics with these procedures.
+
+```fortran
+rank = h%ndims(dataset_name)
+
+character(*), intent(in) :: dataset_name
+```
+
+```fortran
+call h%shape(dataset_name, dims, ierr)
+character(*), intent(in) :: dataset_name
+integer(HSIZE_T), intent(out), allocatable :: dims(:)
+integer, intent(out), optional :: ierr
+```
+
+```fortran
+exists = h%exist(dname)
+!! does dataset "dname" exist in this file
+character(*), intent(in) :: dname
+```
+
+```fortran
+tf = h%is_contig(dname)
+!! is dataset contiguous
+character(*), intent(in) :: dname
+```
+
+```fortran
+tf = h%is_compact(dname)
+!! is dataset compact layout
+character(*), intent(in) :: dname
+```
+
+```fortran
+tf = h%is_chunked(dname)
+!! is dataset chunked
+character(*), intent(in) :: dname
+```
+
+```fortran
+use h5fortran, only: is_hdf5
+
+tf = is_hdf5('myfile.txt')  !< probably false
+tf = is_hdf5('myfile.h5')  !< true if a valid HDF5 file
+```
+
+These are more advanced inquiries into the memory layout of the dataset, for advanced users:
+
+```fortran
+Layout = h%layout(dname)
+!! integer :: H5D_CONTIGUOUS_F, H5D_CHUNKED_F, H5D_VIRTUAL_F, H5D_COMPACT_F
+character(*), intent(in) :: dname
+```
+
+```fortran
+call h%chunks(dname, chunk_size)
+character(*), intent(in) :: dname
+integer(hsize_t), intent(out) :: chunk_size(:)
+```
+
+## create dataset softlink
+
+One of the key features of HDF5 is the ability to create dataset softlinks within an HDF5 file:
+
+```fortran
+call h%softlink(target, link)
+character(*), intent(in) :: target, &  !< target path to link dataset
+                            link  !< soft link path to create
+```
+
+## file write operations
+
+```fortran
+call h%write(dname,value, ierr, chunk_size, istart, iend, stride, compact)
+!! write 0d..7d dataset
+character(*), intent(in) :: dname
+class(*), intent(in) :: value(:)  !< array to write
+integer, intent(in), optional :: chunk_size(1)
+integer, intent(in), optional, dimension(:) :: istart, iend, stride  !< array slicing
+logical, intent(in), optional :: compact  !< faster I/O for sub-64 kB datasets
+integer, intent(out), optional :: ierr  !< 0 if OK
+```
+
+Write dataset attribute (e.g. units or instrument)
+
+```fortran
+call h%writeattr(dname, attr, attrval, ierr)
+character(*), intent(in) :: dname, attr  !< dataset name, attribute name
+class(*), intent(in) :: attrval(:)  !< character, real, integer
+```
+
+## file read operations
+
+Read data from disk to memory
+
+```fortran
+call h%read(dname, value, ierr, istart, iend, stride)
+character(*), intent(in)         :: dname
[…]
