### 2021-03-22 geospace-code/h5fortran#13 @scivision
https://github.com/geospace-code/h5fortran/blob/main/API.md was created, thanks
