+++ .github/workflows/ci_windows.yml
@@ -2,6 +2,7 @@ name: ci_windows
 env:
   CMAKE_GENERATOR: Ninja
   CTEST_SITE: GitHub_Windows
+  CTEST_PARALLEL_LEVEL: 2
 
 on:
   push:
