+++ .github/workflows/ci.yml
@@ -1,5 +1,8 @@
 name: ci
 
+env:
+  CTEST_PARALLEL_LEVEL: 3
+
 on:
   push:
     paths:
