+++ .github/workflows/ci_windows.yml
@@ -85,6 +85,6 @@ jobs:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
       with:
         upload_url: ${{ github.event.release.upload_url }}
-        asset_path: build/package/h5fortran-${{ steps.get_version.outputs.VERSION }}-win64.zip
-        asset_name:  h5fortran-${{ steps.get_version.outputs.VERSION }}-win64.zip
+        asset_path: build/package/h5fortran-${{ steps.get_version.outputs.VERSION }}-win64.tar.zst
+        asset_name:  h5fortran-${{ steps.get_version.outputs.VERSION }}-win64.tar.zst
         asset_content_type: application/zip
