### 2021-04-12 840cb89
cmake: depend conf on template change
