+++ .github/workflows/ci_windows.yml
@@ -49,13 +49,10 @@ jobs:
 
     - uses: actions/checkout@v2
 
-    - name: CDash
-      run: ctest -S ci.cmake -V
-
-    - name: Install package
-      run: |
-        cmake -B build -DCMAKE_INSTALL_PREFIX=~/h5fortran
-        cmake --install build
+    - run: cmake --preset makewin -DCMAKE_INSTALL_PREFIX=~/h5fortran
+    - run: cmake --build --preset default
+    - run: ctest --preset default
+    - run: cmake --install build
 
     - run: cmake -S Examples -B Examples/build -Dh5fortran_ROOT=~/h5fortran
     - run: cmake --build Examples/build --parallel
