### 2021-04-12 845aa03
ci_windows: specify example generator ninja
