+++ fpm.toml
@@ -1,10 +1,10 @@
 name = "h5fortran"
-version = "3.4.3"
-license = "BSD 3-Clause"
-author = "Michael Hirsch"
-copyright = "2020 Michael Hirsch"
-keywords = ["hdf5"]
-homepage = "https://github.com/geospace-code/h5fortran"
+description = "Lightweight object-oriented HDF5 interface"
+categories = "io"
+version = "3.6.6"
 
 [build]
 link = ["z", "hdf5"]
+
+[test]
+source-dir = "src/tests"
