### 2021-04-16 geospace-code/h5fortran#15 @milancurcic
Okay, thanks I found it, although it's quite buried at the bottom of the CMake section. Can you place it in its own section so its more visible? I don't know what's a good section name. "Non-CMake"? "Other build systems"?
