### 2021-04-16 geospace-code/h5fortran#20 @milancurcic
Build instructions with CMake not working
From README.md:

> cmake --preset=ninja
> cmake --build build

Here's what I get:

```
$ cmake --preset=ninja
CMake Error: The source directory "/home/milan/Work/fortran/h5fortran/--preset=ninja" does not exist.
Specify --help for usage, or press the help button on the CMake GUI.
```

and

```
$ cmake --build build
Error: /home/milan/Work/fortran/h5fortran/build is not a directory
```

Should this rather be `cmake -B build`, which works for me?

My CMake is 3.18.4.

(working on openjournals/joss-reviews#2842)
