### 2021-04-16 geospace-code/h5fortran#21 @milancurcic
Running CTest attempts to open MATLAB
Running CTest attempts to open MATLAB. The first time I thought it was coincidental, but I can reproduce it: It happens on running the `h5fortran:check_shape` test.

![h5fortran-matlab](https://user-images.githubusercontent.com/4133310/115048982-2f49eb80-9ea8-11eb-8e18-2d65512450e3.png)

The test hangs until I cancel out of the MATLAB pop-up, after which it fails with a timeout.

```
...
      Start 18: h5fortran:check_shape
18/18 Test #18: h5fortran:check_shape ...............***Timeout  90.09 sec

83% tests passed, 3 tests failed out of 18

Label Time Summary:
core     =   0.01 sec*proc (1 test)
shaky    =  90.59 sec*proc (6 tests)
unit     =   0.41 sec*proc (11 tests)

Total Test time (real) =  91.04 sec

The following tests FAILED:
	 13 - h5fortran:fail_read_size_mismatch (SEGFAULT)
	 14 - h5fortran:fail_read_rank_mismatch (SEGFAULT)
	 18 - h5fortran:check_shape (Timeout)
Errors while running CTest
```

**System details**

* Ubuntu 18.10
* HDF5 library version: 1.10.0-patch1+docs-4build2
* gcc-8.3.0

(working on openjournals/joss-reviews#2842)
