### 2021-04-16 geospace-code/h5fortran#22 @milancurcic
Tests 13 and 14 segfault on gcc-9.3.0, hdf5-1.10.6 on Ubuntu 18.10
```
$ ctest
Test project /home/milan/Work/fortran/h5fortran/build
      Start  1: h5fortran:minimal
 1/18 Test  #1: h5fortran:minimal ...................   Passed    0.01 sec
      Start  2: h5fortran:array
 2/18 Test  #2: h5fortran:array .....................   Passed    0.02 sec
      Start  3: h5fortran:attributes
 3/18 Test  #3: h5fortran:attributes ................   Passed    0.01 sec
      Start  4: h5fortran:deflate
 4/18 Test  #4: h5fortran:deflate ...................   Passed    0.27 sec
      Start  5: h5fortran:error
 5/18 Test  #5: h5fortran:error .....................   Passed    0.01 sec
      Start  6: h5fortran:exist
 6/18 Test  #6: h5fortran:exist .....................   Passed    0.01 sec
      Start  7: h5fortran:module
 7/18 Test  #7: h5fortran:module ....................   Passed    0.01 sec
      Start  8: h5fortran:layout
 8/18 Test  #8: h5fortran:layout ....................   Passed    0.01 sec
      Start  9: h5fortran:lt
 9/18 Test  #9: h5fortran:lt ........................   Passed    0.03 sec
      Start 10: h5fortran:scalar
10/18 Test #10: h5fortran:scalar ....................   Passed    0.02 sec
      Start 11: h5fortran:shape
11/18 Test #11: h5fortran:shape .....................   Passed    0.01 sec
      Start 12: h5fortran:string
12/18 Test #12: h5fortran:string ....................   Passed    0.01 sec
      Start 13: h5fortran:fail_read_size_mismatch
13/18 Test #13: h5fortran:fail_read_size_mismatch ...***Exception: SegFault  0.22 sec
      Start 14: h5fortran:fail_read_rank_mismatch
14/18 Test #14: h5fortran:fail_read_rank_mismatch ...***Exception: SegFault  0.20 sec
      Start 15: h5fortran:fail_nonexist_variable
15/18 Test #15: h5fortran:fail_nonexist_variable ....   Passed    0.03 sec
      Start 16: h5fortran:fail_unknown_read
16/18 Test #16: h5fortran:fail_unknown_read .........   Passed    0.03 sec
      Start 17: h5fortran:fail_unknown_write
17/18 Test #17: h5fortran:fail_unknown_write ........   Passed    0.03 sec
      Start 18: h5fortran:check_shape
18/18 Test #18: h5fortran:check_shape ...............***Timeout  90.09 sec

83% tests passed, 3 tests failed out of 18

Label Time Summary:
core     =   0.01 sec*proc (1 test)
shaky    =  90.59 sec*proc (6 tests)
unit     =   0.41 sec*proc (11 tests)

Total Test time (real) =  91.04 sec

The following tests FAILED:
	 13 - h5fortran:fail_read_size_mismatch (SEGFAULT)
	 14 - h5fortran:fail_read_rank_mismatch (SEGFAULT)
	 18 - h5fortran:check_shape (Timeout)
Errors while running CTest
```

**System details**

Ubuntu 18.10
HDF5 library version: 1.10.0-patch1+docs-4build2
gcc-8.3.0

Let me know if and how to look deeper into this.

Test 18 failure is reported in #21.

(working on openjournals/joss-reviews#2842)
