### 2021-04-19 geospace-code/h5fortran#21 @scivision
this was intended for developers, and I made it not run by default now
