+++ cmake/options.cmake
@@ -2,11 +2,14 @@
 option(hdf5_external "Build HDF5 library")
 option(dev "developer mode")
 
+option(matlab "check HDF5 file writes with Matlab" off)
 option(concepts "conceptual testing, for devs only" off)
 
 set(CMAKE_EXPORT_COMPILE_COMMANDS on)
 
-if(NOT dev)
+if(dev)
+
+else()
   set(EP_UPDATE_DISCONNECTED true)
 endif()
 
