### 2021-04-19 geospace-code/h5fortran#22 @scivision
this is another case where tests that are only for devs are now off by default. These tests try to crash HDF5 library, and sometimes segfault. This is a known issue with CMake that WILL_FAIL with segfault is still an error with CMake. So it's meant to used on manual human-supervised tests e.g. by devs.
