### 2021-04-19 geospace-code/h5fortran#20 @scivision
I updated the docs--presets assume CMake >= 3.19. So the docs now work with Cmake < 3.19 as well
