### 2021-04-19 geospace-code/h5fortran#15 @scivision
yes I made a "compiler wrapper" section for this
