### 2021-04-19 geospace-code/h5fortran#20 @milancurcic
Thanks! Should [this section](https://github.com/geospace-code/h5fortran/blob/main/Install.md#cmake) be updated as well? At least with a note that --preset needs CMake >= 3.19.
