### 2021-04-19 7f877c6
doc: cmake w/o preset
