### 2021-04-19 geospace-code/h5fortran#20 @scivision
Thanks I updated that too
