### 2021-04-26 #2842 @pdebuyl
Thanks @bast for your review! Can you go over the checklist? It is part of the review and I can't proceed until all is checked :-)

@scivision I will start to process the "editorial steps" https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance already.
