### 2021-04-29 359a563
better install / externalProject defaults
