+++ cmake/build_zlib.cmake
@@ -9,7 +9,7 @@ else()
 endif()
 
 if(NOT DEFINED ZLIB_ROOT)
-  set(ZLIB_ROOT ${PROJECT_BINARY_DIR}/zlib)
+  set(ZLIB_ROOT ${CMAKE_INSTALL_PREFIX})
 endif()
 
 set(ZLIB_INCLUDE_DIR ${ZLIB_ROOT}/include)
@@ -18,7 +18,7 @@ set(ZLIB_LIBRARY ${ZLIB_ROOT}/lib/${ZLIB_name})
 ExternalProject_Add(ZLIB
 URL ${zlib_url}
 URL_HASH SHA1=${zlib_sha1}
-TLS_VERIFY ON
+UPDATE_DISCONNECTED ${EP_UPDATE_DISCONNECTED}
 CONFIGURE_HANDLED_BY_BUILD ON
 INACTIVITY_TIMEOUT 30
 CMAKE_ARGS -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX:PATH=${ZLIB_ROOT}
