+++ cmake/build_hdf5.cmake
@@ -10,7 +10,7 @@ set(HDF5_VERSION 1.10.7)
 include(ExternalProject)
 
 if(NOT DEFINED HDF5_ROOT)
-  set(HDF5_ROOT ${PROJECT_BINARY_DIR}/hdf5)
+  set(HDF5_ROOT ${CMAKE_INSTALL_PREFIX})
 endif()
 
 set(HDF5_LIBRARIES)
@@ -23,7 +23,7 @@ set(HDF5_INCLUDE_DIRS ${HDF5_ROOT}/include)
 # --- Zlib
 set(zlib_root -DHDF5_ENABLE_Z_LIB_SUPPORT:BOOL=ON -DZLIB_USE_EXTERNAL:BOOL=OFF)
 
-include(${CMAKE_CURRENT_LIST_DIR}/zlib_setup.cmake)
+include(${CMAKE_CURRENT_LIST_DIR}/build_zlib.cmake)
 
 # --- HDF5
 # https://forum.hdfgroup.org/t/issues-when-using-hdf5-as-a-git-submodule-and-using-cmake-with-add-subdirectory/7189/2
@@ -31,7 +31,7 @@ include(${CMAKE_CURRENT_LIST_DIR}/zlib_setup.cmake)
 ExternalProject_Add(HDF5
 URL ${hdf5_url}
 URL_HASH SHA1=${hdf5_sha1}
-TLS_VERIFY ON
+UPDATE_DISCONNECTED ${EP_UPDATE_DISCONNECTED}
 CONFIGURE_HANDLED_BY_BUILD ON
 INACTIVITY_TIMEOUT 30
 CMAKE_ARGS ${zlib_root} -DCMAKE_INSTALL_PREFIX:PATH=${HDF5_ROOT} -DHDF5_GENERATE_HEADERS:BOOL=false -DHDF5_DISABLE_COMPILER_WARNINGS:BOOL=true -DBUILD_SHARED_LIBS:BOOL=false -DCMAKE_BUILD_TYPE=Release -DHDF5_BUILD_FORTRAN:BOOL=true -DHDF5_BUILD_CPP_LIB:BOOL=false -DHDF5_BUILD_TOOLS:BOOL=false -DBUILD_TESTING:BOOL=false -DHDF5_BUILD_EXAMPLES:BOOL=false
