+++ cmake/Config.cmake.in
@@ -5,6 +5,6 @@ include(CMakeFindDependencyMacro)
 list(APPEND CMAKE_MODULE_PATH ${CMAKE_CURRENT_LIST_DIR}/../Modules)
 find_dependency(HDF5 COMPONENTS Fortran)
 
-include("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@Targets.cmake")
+include(${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@Targets.cmake)
 
 check_required_components(@PROJECT_NAME@)
