+++ cmake/build_zlib.cmake
@@ -17,7 +17,7 @@ set(ZLIB_LIBRARY ${ZLIB_ROOT}/lib/${ZLIB_name})
 
 ExternalProject_Add(ZLIB
 URL ${zlib_url}
-URL_HASH SHA1=${zlib_sha1}
+URL_HASH SHA256=${zlib_sha256}
 UPDATE_DISCONNECTED ${EP_UPDATE_DISCONNECTED}
 CONFIGURE_HANDLED_BY_BUILD ON
 INACTIVITY_TIMEOUT 30
