+++ cmake/build_hdf5.cmake
@@ -30,7 +30,7 @@ include(${CMAKE_CURRENT_LIST_DIR}/build_zlib.cmake)
 
 ExternalProject_Add(HDF5
 URL ${hdf5_url}
-URL_HASH SHA1=${hdf5_sha1}
+URL_HASH SHA256=${hdf5_sha256}
 UPDATE_DISCONNECTED ${EP_UPDATE_DISCONNECTED}
 CONFIGURE_HANDLED_BY_BUILD ON
 INACTIVITY_TIMEOUT 30
