### 2021-05-07 5540f9b
find_package(HDF5 ... HL) to help keep factory compatibility
