+++ cmake/compilers.cmake
@@ -18,16 +18,16 @@ endif()
 if(CMAKE_Fortran_COMPILER_ID STREQUAL Intel)
   if(WIN32)
     # add_compile_options(/Qdiag-error-limit:3)
-    string(APPEND CMAKE_Fortran_FLAGS " /traceback /warn /heap-arrays")
-    string(APPEND CMAKE_Fortran_FLAGS_DEBUG " /check:bounds /debug:all")
+    string(APPEND CMAKE_Fortran_FLAGS " /warn /heap-arrays")
+    string(APPEND CMAKE_Fortran_FLAGS_DEBUG " /traceback /check:bounds /debug:all")
   else()
     # add_compile_options(-diag-error-limit=3)
-    string(APPEND CMAKE_Fortran_FLAGS " -traceback -warn -heap-arrays")
-    string(APPEND CMAKE_Fortran_FLAGS_DEBUG " -check all -debug extended")
+    string(APPEND CMAKE_Fortran_FLAGS " -warn -heap-arrays")
+    string(APPEND CMAKE_Fortran_FLAGS_DEBUG " -traceback -check all -debug extended")
   endif()
 elseif(CMAKE_Fortran_COMPILER_ID STREQUAL GNU)
   string(APPEND CMAKE_Fortran_FLAGS " -mtune=native -Wall -fimplicit-none")
-  string(APPEND CMAKE_Fortran_FLAGS_RELEASE " -Wno-maybe-uninitialized")
+  string(APPEND CMAKE_Fortran_FLAGS_RELEASE " -fno-backtrace -Wno-maybe-uninitialized")
   string(APPEND CMAKE_Fortran_FLAGS_RELWITHDEBINFO " -Wno-maybe-uninitialized")
   string(APPEND CMAKE_Fortran_FLAGS_DEBUG " -Wextra -fcheck=all -Werror=array-bounds")
 elseif(CMAKE_Fortran_COMPILER_ID STREQUAL NAG)
