### 2021-05-07 a8d1c17
better install location default
