+++ cmake/build_zlib.cmake
@@ -9,7 +9,11 @@ else()
 endif()
 
 if(NOT DEFINED ZLIB_ROOT)
-  set(ZLIB_ROOT ${CMAKE_INSTALL_PREFIX})
+  if(CMAKE_INSTALL_PREFIX_INITIALIZED_TO_DEFAULT)
+    set(ZLIB_ROOT ${PROJECT_BINARY_DIR})
+  else()
+    set(ZLIB_ROOT ${CMAKE_INSTALL_PREFIX})
+  endif()
 endif()
 
 set(ZLIB_INCLUDE_DIR ${ZLIB_ROOT}/include)
