+++ cmake/build_hdf5.cmake
@@ -10,7 +10,11 @@ set(HDF5_VERSION 1.10.7)
 include(ExternalProject)
 
 if(NOT DEFINED HDF5_ROOT)
-  set(HDF5_ROOT ${CMAKE_INSTALL_PREFIX})
+  if(CMAKE_INSTALL_PREFIX_INITIALIZED_TO_DEFAULT)
+    set(HDF5_ROOT ${PROJECT_BINARY_DIR})
+  else()
+    set(HDF5_ROOT ${CMAKE_INSTALL_PREFIX})
+  endif()
 endif()
 
 set(HDF5_LIBRARIES)
