### 2021-05-07 7cdabb7
ensure _ROOT is non-empty for externalProject
