+++ cmake/build_zlib.cmake
@@ -3,7 +3,7 @@
 include(ExternalProject)
 
 if(WIN32)
-  set(ZLIB_name ${CMAKE_STATIC_LIBRARY_PREFIX}zlibstatic${CMAKE_STATIC_LIBRARY_SUFFIX})
+  set(ZLIB_name ${CMAKE_STATIC_LIBRARY_PREFIX}zlib${CMAKE_STATIC_LIBRARY_SUFFIX})
 else()
   set(ZLIB_name ${CMAKE_STATIC_LIBRARY_PREFIX}z${CMAKE_STATIC_LIBRARY_SUFFIX})
 endif()
@@ -23,10 +23,9 @@ set(ZLIB_LIBRARY ${ZLIB_ROOT}/lib/${ZLIB_name})
 ExternalProject_Add(ZLIB
 URL ${zlib_url}
 URL_HASH SHA256=${zlib_sha256}
-UPDATE_DISCONNECTED ${EP_UPDATE_DISCONNECTED}
 CONFIGURE_HANDLED_BY_BUILD ON
 INACTIVITY_TIMEOUT 30
-CMAKE_ARGS -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX:PATH=${ZLIB_ROOT}
+CMAKE_ARGS -DZLIB_COMPAT:BOOL=on -DZLIB_ENABLE_TESTS:BOOL=off -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX:PATH=${ZLIB_ROOT}
 BUILD_BYPRODUCTS ${ZLIB_LIBRARY}
 )
 
