### 2021-05-10 d07d465
zlib-ng + Ninja + Intel Windows needs BUILD_SHARED_LIBS=off
Intel Windows needs Ninja generator to build HDF5, but Intel Windows
with Zlib and Ninja will see a duplicate static/shared target.
We only use static Zlib so do this.
