+++ cmake/build_zlib.cmake
@@ -25,7 +25,7 @@ URL ${zlib_url}
 URL_HASH SHA256=${zlib_sha256}
 CONFIGURE_HANDLED_BY_BUILD ON
 INACTIVITY_TIMEOUT 30
-CMAKE_ARGS -DZLIB_COMPAT:BOOL=on -DZLIB_ENABLE_TESTS:BOOL=off -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX:PATH=${ZLIB_ROOT}
+CMAKE_ARGS -DZLIB_COMPAT:BOOL=on -DZLIB_ENABLE_TESTS:BOOL=off -DBUILD_SHARED_LIBS:BOOL=off -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX:PATH=${ZLIB_ROOT}
 BUILD_BYPRODUCTS ${ZLIB_LIBRARY}
 )
 
