### 2021-05-12 8177888
add -Dzlib_legacy:BOOL option for Zlib 1.x fallback (default 2.x)
