+++ cmake/build_zlib.cmake
@@ -3,7 +3,11 @@
 include(ExternalProject)
 
 if(WIN32)
-  set(ZLIB_name ${CMAKE_STATIC_LIBRARY_PREFIX}zlib${CMAKE_STATIC_LIBRARY_SUFFIX})
+  if(zlib_legacy)
+    set(ZLIB_name ${CMAKE_STATIC_LIBRARY_PREFIX}zlibstatic${CMAKE_STATIC_LIBRARY_SUFFIX})
+  else()
+    set(ZLIB_name ${CMAKE_STATIC_LIBRARY_PREFIX}zlib${CMAKE_STATIC_LIBRARY_SUFFIX})
+  endif()
 else()
   set(ZLIB_name ${CMAKE_STATIC_LIBRARY_PREFIX}z${CMAKE_STATIC_LIBRARY_SUFFIX})
 endif()
@@ -24,7 +28,7 @@ ExternalProject_Add(ZLIB
 URL ${zlib_url}
 URL_HASH SHA256=${zlib_sha256}
 CONFIGURE_HANDLED_BY_BUILD ON
-INACTIVITY_TIMEOUT 30
+INACTIVITY_TIMEOUT 15
 CMAKE_ARGS -DZLIB_COMPAT:BOOL=on -DZLIB_ENABLE_TESTS:BOOL=off -DBUILD_SHARED_LIBS:BOOL=off -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX:PATH=${ZLIB_ROOT}
 BUILD_BYPRODUCTS ${ZLIB_LIBRARY}
 )
