### 2021-05-13 #2842 @pdebuyl
Hi @scivision

The documentation at https://geospace-code.github.io/h5fortran/ lacks a "narrative documentation", what is typically called the user guide. The README of the project contains most of what's necessary though, could you integrate the README to the FORD documentation output, with install.md and examples.md ? This way, the rendered docs would be complete.

Regarding installation, there is a bit of duplication between the README and the install page. This is fine, but if you look only at install.md you don't see that h5fortran can install an extra instance of HDF5, which can be annoying for many users (HPC users, for instance).

In the README, you mention the lack of support for complex data. Could you mention logical as well? And include both in the article?

Do you think that a mention of [h5xx](https://github.com/fhoefling/h5xx) would be relevant? I don't use it myself but it seems similar in scope: offering a simplified HDF5 API for typical objects (h5xx is based on templates and inspired by h5py).

If there are other research activities that rely on h5fortran, you are welcome to mention them as well.

Finally, I don't see the use case of appending to extensible datasets mentioned. Is it supported by h5fortran? Are there specific routines to ease that process for users? If not, it should be mentioned as a limitation as it is very typical for simulation work.

Sorry for the delay here, Pierre.
