+++ src/tests/test_shape.f90
@@ -1,5 +1,5 @@
 program test_shape
-!! This program shows how HDF5 dimension orders are distinct in different langauges
+!! This program shows how HDF5 dimension orders are distinct in different languages
 use h5fortran, only: hdf5_file,hsize_t, is_hdf5
 use, intrinsic:: iso_fortran_env, only: real64, stdout=>output_unit, stderr=>error_unit
 
