### 2021-05-20 3530cb9
zlib: use patched 2.0.3 that works with GCC MPICC
