+++ cmake/build_zlib.cmake
@@ -24,6 +24,16 @@ endif()
 set(ZLIB_INCLUDE_DIR ${ZLIB_ROOT}/include)
 set(ZLIB_LIBRARY ${ZLIB_ROOT}/lib/${ZLIB_name})
 
+if(zlib_git)
+ExternalProject_Add(ZLIB
+GIT_REPOSITORY ${zlib_git}
+GIT_TAG ${zlib_tag}
+CONFIGURE_HANDLED_BY_BUILD ON
+INACTIVITY_TIMEOUT 15
+CMAKE_ARGS -DZLIB_COMPAT:BOOL=on -DZLIB_ENABLE_TESTS:BOOL=off -DBUILD_SHARED_LIBS:BOOL=off -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX:PATH=${ZLIB_ROOT}
+BUILD_BYPRODUCTS ${ZLIB_LIBRARY}
+)
+else()
 ExternalProject_Add(ZLIB
 URL ${zlib_url}
 URL_HASH SHA256=${zlib_sha256}
@@ -32,7 +42,7 @@ INACTIVITY_TIMEOUT 15
 CMAKE_ARGS -DZLIB_COMPAT:BOOL=on -DZLIB_ENABLE_TESTS:BOOL=off -DBUILD_SHARED_LIBS:BOOL=off -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX:PATH=${ZLIB_ROOT}
 BUILD_BYPRODUCTS ${ZLIB_LIBRARY}
 )
-
+endif(zlib_git)
 
 # --- imported target
 
