+++ cmake/libraries.cmake
@@ -1,6 +1,6 @@
 if(CMAKE_VERSION VERSION_LESS 3.19)
-  set(zlib_url https://github.com/zlib-ng/zlib-ng/archive/refs/tags/2.0.3.zip)
-  set(zlib_sha256 235232f622d17cb4070f67ebeb06b0df8cf0894f9045ea6c51782479318841d2)
+  set(zlib_git https://github.com/zlib-ng/zlib-ng.git)
+  set(zlib_tag bde11b12775aed5113040de319f484da879a3f04)
 
   set(hdf5_url https://github.com/HDFGroup/hdf5/archive/hdf5-1_10_7.zip)
   set(hdf5_sha256 89ea0b117c27a8f7587e601b65fa1427b28db3a6ee66ffff0178509514128f18)
@@ -13,8 +13,8 @@ else()
     string(JSON zlib_url GET ${_libj} zlib1 url)
     string(JSON zlib_sha256 GET ${_libj} zlib1 sha256)
   else()
-    string(JSON zlib_url GET ${_libj} zlib2 url)
-    string(JSON zlib_sha256 GET ${_libj} zlib2 sha256)
+    string(JSON zlib_git GET ${_libj} zlib2 git)
+    string(JSON zlib_tag GET ${_libj} zlib2 tag)
   endif()
 
   string(JSON hdf5_url GET ${_libj} hdf5 url)
