### 2021-05-20 9a8fdbd
write: more detailed error stop messages
