+++ src/write/write.f90
@@ -22,16 +22,13 @@ integer :: ierr
 integer(HID_T) :: pid, space_id, ds_id
 integer(HSIZE_T) :: vdims(size(dims))
 
-if(.not.self%is_open) error stop 'h5fortran:write: file handle is not open'
+if(.not.self%is_open) error stop 'h5fortran:write: file handle is not open: ' // self%filename
 
 call h5ltpath_valid_f(self%lid, dname, .true., exists, ierr)
 !! h5lexists_f can false error with groups--just use h5ltpath_valid
 
 !! stricter than self%exists() since we're creating and/or writing variable
-if (ierr /= 0) then
-  write(stderr,*) 'ERROR:h5fortran:create: variable path invalid: ',dname, ' in ', self%filename
-  error stop 6
-endif
+if (ierr /= 0) error stop 'ERROR:h5fortran:create: variable path invalid: ' // dname // ' in ' // self%filename
 
 !> allow user to specify int4 or int8 dims
 select type (dims)
@@ -40,8 +37,7 @@ type is (integer(int32))
 type is (integer(hsize_t))
   vdims = dims
 class default
-  write(stderr,*) 'ERROR:h5fortran:create: wrong type for dims: ', dname, self%filename
-  error stop 5
+  error stop 'ERROR:h5fortran:create: wrong type for dims: ' // dname // " " // self%filename
 end select
 
 if(self%debug) print *,'h5fortran:TRACE:create:exists: ' // dname, exists
@@ -51,16 +47,13 @@ if(exists) then
   !! FIXME: read and write slice shape not checked; but should check in future versions
   !> open dataset
   call h5dopen_f(self%lid, dname, ds_id, ierr)
-  if (ierr /= 0) then
-    write(stderr,*) 'ERROR:h5fortran:create: could not open ', dname, ' in ', self%filename
-    error stop
-  endif
+  if (ierr /= 0) error stop 'ERROR:h5fortran:create: could not open ' // dname // ' in ' // self%filename
 
   if(present(did)) did = ds_id
   if(present(sid)) then
     call h5dget_space_f(ds_id, sid, ierr)
-    if(ierr /= 0) error stop 'h5fortran:create could not get dataset'
-  endif
+    if(ierr /= 0) error stop 'h5fortran:create could not get dataset ' // dname // ' in ' // self%filename
+  end if
   return
 endif
 
@@ -69,10 +62,8 @@ if(self%debug) print *, 'h5fortran:TRACE1: ' // dname
 !> Only new datasets go past this point
 
 call self%write_group(dname, ierr)
-if (ierr /= 0) then
-  write(stderr,*) 'ERROR:h5fortran:create: could not create group for ', dname, ' in ', self%filename
-  error stop
-endif
+if (ierr /= 0) error stop 'ERROR:h5fortran:create: could not create group for ' // dname // ' in ' // self%filename
+
 
 !> create properties
 pid = 0 !< sentinel
@@ -80,7 +71,7 @@ pid = 0 !< sentinel
 if(size(vdims) >= 2) then
   if(self%debug) print *, 'h5fortran:TRACE:create: deflate: ' // dname
   call set_deflate(self, vdims, pid, ierr, chunk_size)
-  if (ierr/=0) error stop 'ERROR:h5fortran:create: problem setting deflate on'
+  if (ierr/=0) error stop 'ERROR:h5fortran:create: problem setting deflate on ' // dname
 endif
 
 if(present(compact)) then
@@ -89,10 +80,10 @@ if(present(compact)) then
 if(compact .and. pid == 0 .and. product(vdims) * 8 < 60000)  then
[…]
