### 2021-05-20 f4f5d1d
hdf_create: make dims hsize_t to avoid ambiguous specifications
