+++ src/write/write.f90
@@ -20,7 +20,6 @@ module procedure hdf_create
 logical :: exists
 integer :: ierr
 integer(HID_T) :: pid, space_id, ds_id
-integer(HSIZE_T) :: vdims(size(dims))
 
 if(.not.self%is_open) error stop 'h5fortran:write: file handle is not open: ' // self%filename
 
@@ -30,20 +29,10 @@ call h5ltpath_valid_f(self%lid, dname, .true., exists, ierr)
 !! stricter than self%exists() since we're creating and/or writing variable
 if (ierr /= 0) error stop 'ERROR:h5fortran:create: variable path invalid: ' // dname // ' in ' // self%filename
 
-!> allow user to specify int4 or int8 dims
-select type (dims)
-type is (integer(int32))
-  vdims = int(dims, int64)
-type is (integer(hsize_t))
-  vdims = dims
-class default
-  error stop 'ERROR:h5fortran:create: wrong type for dims: ' // dname // " " // self%filename
-end select
-
 if(self%debug) print *,'h5fortran:TRACE:create:exists: ' // dname, exists
 
 if(exists) then
-  if (.not.present(istart)) call hdf_shape_check(self, dname, vdims)
+  if (.not.present(istart)) call hdf_shape_check(self, dname, dims)
   !! FIXME: read and write slice shape not checked; but should check in future versions
   !> open dataset
   call h5dopen_f(self%lid, dname, ds_id, ierr)
@@ -68,16 +57,16 @@ if (ierr /= 0) error stop 'ERROR:h5fortran:create: could not create group for ' 
 !> create properties
 pid = 0 !< sentinel
 
-if(size(vdims) >= 2) then
+if(size(dims) >= 2) then
   if(self%debug) print *, 'h5fortran:TRACE:create: deflate: ' // dname
-  call set_deflate(self, vdims, pid, ierr, chunk_size)
+  call set_deflate(self, dims, pid, ierr, chunk_size)
   if (ierr/=0) error stop 'ERROR:h5fortran:create: problem setting deflate on ' // dname
 endif
 
 if(present(compact)) then
 ! print *, "TRACE1: COMPACT", compact
 !! don't set COMPACT after CHUNKED, will fail. And it's either or anyway.
-if(compact .and. pid == 0 .and. product(vdims) * 8 < 60000)  then
+if(compact .and. pid == 0 .and. product(dims) * 8 < 60000)  then
 !! 64000 byte limit, here we assumed 8 bytes / element
   call h5pcreate_f(H5P_DATASET_CREATE_F, pid, ierr)
   if (check(ierr, self%filename)) error stop "h5fortran:h5pcreate: " // dname
@@ -88,10 +77,10 @@ endif
 endif
 
 !> create dataspace
-if(size(vdims) == 0) then
+if(size(dims) == 0) then
   call h5screate_f(H5S_SCALAR_F, space_id, ierr)
 else
-  call h5screate_simple_f(size(vdims), vdims, space_id, ierr)
+  call h5screate_simple_f(size(dims), dims, space_id, ierr)
 endif
 if (check(ierr, self%filename, dname)) error stop "h5fortran:h5screate: " // dname
 
