### 2021-05-20 621d6af
write_scalar: use TKR distinct instead of "select type"
This helps runtime robustness/speed as well as avoiding bug in
GCC 11.1 causing runtime crash
