+++ src/write/writer.in.f90
@@ -7,50 +7,110 @@ implicit none (type, external)
 contains
 
 
-module procedure hdf_write_scalar
+module procedure hdf_write_scalar_r32
 
 integer(HID_T)  :: sid, did
 integer(HSIZE_T), allocatable :: dims(:)
 integer :: ier
 
-if(.not.self%is_open) error stop 'h5fortran:writer: file handle is not open'
+if(.not.self%is_open) error stop 'h5fortran:write: file handle is not open'
 
 allocate(dims(0))
 
-did = 0
-sid = 0
-!! sentinel
+call hdf_create(self, dname, H5T_NATIVE_REAL, dims, sid, did, compact=compact)
+call h5dwrite_f(did, H5T_NATIVE_REAL, value, dims, ier)
 
-select type (value)
-type is (character(*))
-  call h5ltmake_dataset_string_f(self%lid, dname, value, ier)
-type is (real(real64))
-  !! NOTE: 0d does not use chunk_size
-  call hdf_create(self,dname, H5T_NATIVE_DOUBLE, dims,sid,did, compact=compact)
-  call h5dwrite_f(did, H5T_NATIVE_DOUBLE, value, dims, ier)
-type is (real(real32))
-  call hdf_create(self,dname, H5T_NATIVE_REAL, dims,sid,did, compact=compact)
-  call h5dwrite_f(did, H5T_NATIVE_REAL, value, dims, ier)
-type is (integer(int32))
-  call hdf_create(self,dname, H5T_NATIVE_INTEGER, dims,sid,did, compact=compact)
-  call h5dwrite_f(did, H5T_NATIVE_INTEGER, value, dims, ier)
-type is (integer(int64))
-  call hdf_create(self,dname, H5T_STD_I64LE, dims,sid,did, compact=compact)
-  call h5dwrite_f(did, H5T_STD_I64LE, value, dims, ier)
-class default
-  error stop 'h5fortran: invalid data type'
-end select
-if (ier/=0) then
-  write(stderr,*) 'h5fortran:ERROR: could not write ',dname, ' to ', self%filename
-  error stop
-endif
+if (ier/=0) error stop 'h5fortran:write: could not write ' // dname // ' to ' // self%filename
 
 call hdf_wrapup(did, sid, ier)
 
 if (present(ierr)) ierr = ier
-if (check(ier, self%filename, dname) .and. .not.present(ierr)) error stop
+if (check(ier, self%filename, dname) .and. .not.present(ierr)) error stop "h5fortran:write:wrapup: " // dname
 
-end procedure hdf_write_scalar
+end procedure hdf_write_scalar_r32
+
+
+module procedure hdf_write_scalar_r64
+
+integer(HID_T)  :: sid, did
+integer(HSIZE_T), allocatable :: dims(:)
+integer :: ier
+
+if(.not.self%is_open) error stop 'h5fortran:write: file handle is not open'
+
+allocate(dims(0))
+
+call hdf_create(self, dname, H5T_NATIVE_DOUBLE, dims, sid, did, compact=compact)
+call h5dwrite_f(did, H5T_NATIVE_DOUBLE, value, dims, ier)
+
+if (ier/=0) error stop 'h5fortran:write: could not write ' // dname // ' to ' // self%filename
+
+call hdf_wrapup(did, sid, ier)
+
+if (present(ierr)) ierr = ier
+if (check(ier, self%filename, dname) .and. .not.present(ierr)) error stop "h5fortran:write:wrapup: " // dname
+
+end procedure hdf_write_scalar_r64
+
+
+module procedure hdf_write_scalar_i32
+
+integer(HID_T)  :: sid, did
+integer(HSIZE_T), allocatable :: dims(:)
[…]
