### 2021-05-20 81cb371
add %write_{r32,r64,i32,i64,char} methods for user-specific
this is optional vs. the generic %write, for where user
wants to be explicit about type written
