+++ src/tests/test_string.f90
@@ -18,6 +18,7 @@ character(*), parameter :: path='test_string.h5'
 call h%initialize(path, status='replace')
 
 call h%write('/little', '42')
+call h%write_char('/little_char', '42')
 call h%write('/MySentence', 'this is a little sentence.')
 
 call h%finalize()
