### 2021-05-21 7695ce7
write: TKR for array for stability like scalar
