### 2021-05-21 7fba52d
write:array add %write_{r32,r64,i32,i64,char} methods for user-specific
    this is optional vs. the generic %write, for where user
    wants to be explicit about type written
