### 2021-06-08 0dcb56c
build_hdf5.py: zlib 2.x updates
