### 2021-06-14 6971d8b
zlib 2.0.4, which fixes mpicc failure to detect compiler
