### 2021-06-14 4523aa2
%open, %close is for file. %open_group,%close_group for groups
also add destructor for files

add auto-close test
