+++ src/tests/test_scalar.f90
@@ -29,7 +29,7 @@ r1 = i1
 i1_64 = i1
 
 !> write
-call h%initialize(fn, status='replace')
+call h%open(fn, status='replace')
 !> scalar tests
 call h%write('/scalar_int32', 42_int32)
 call h%write_i32('/scalar_i32', 42_int32)
@@ -58,11 +58,11 @@ print *, 'PASSED: vector write'
 call h%write('scalar_real32', 42.)
 call h%write('scalar_int32', 42_int32)
 call h%write('scalar_int64', 42_int64)
-call h%finalize()
+call h%close()
 
 !> read
 
-call h%initialize(fn, status='old', action='r')
+call h%open(fn, status='old', action='r')
 
 call h%read('/scalar_int32', it)
 call h%read('/scalar_int64', it_64)
@@ -116,6 +116,6 @@ if (.not. h%filename == fn) then
   error stop
 endif
 
-call h%finalize()
+call h%close()
 
 end program
