+++ src/tests/test_fail_read_size_mismatch.f90
@@ -4,9 +4,9 @@ use h5fortran, only : hdf5_file
 implicit none (type, external)
 type(hdf5_file) :: h
 
-call h%initialize('mismatch.h5', status='scratch', verbose=.False.)
+call h%open('mismatch.h5', status='scratch', verbose=.False.)
 call h%write('/int32-1d', [-1,1])
 call h%write('/int32-1d', [-1])
-call h%finalize()
+call h%close()
 
 end program
