+++ src/read/reader_lt_template.in.f90
@@ -1,9 +1,9 @@
 type(hdf5_file) :: h
 integer :: ier
 
-call h%initialize(filename, ier, status='old', action='r')
+call h%open(filename, ier, status='old', action='r')
 if (ier == 0) call h%read(dname, value, ier)
-if (ier == 0) call h%finalize(ier)
+if (ier == 0) call h%close(ier)
 
 if (present(ierr)) ierr = ier
 if (check(ier, filename, dname) .and. .not.present(ierr)) error stop
