+++ src/tests/test_error.f90
@@ -21,7 +21,7 @@ subroutine test_nonexist_old_file()
 integer :: ierr
 type(hdf5_file) :: h
 
-call h%initialize('not-exist.h5', ierr, status='old', action='read', verbose=.false.)
+call h%open('not-exist.h5', ierr, status='old', action='read', verbose=.false.)
 if (ierr==0) error stop 'should have had ierr/=0 on non-existing old file'
 end subroutine test_nonexist_old_file
 
@@ -30,7 +30,7 @@ subroutine test_nonexist_unknown_file()
 integer :: ierr
 type(hdf5_file) :: h
 
-call h%initialize('not-exist.h5', ierr, status='unknown', action='read', verbose=.false.)
+call h%open('not-exist.h5', ierr, status='unknown', action='read', verbose=.false.)
 if (ierr==0) error stop 'should have had ierr/=0 on non-existing unknown read file'
 end subroutine test_nonexist_unknown_file
 
@@ -44,7 +44,7 @@ character(*), parameter :: filename = 'bad.h5'
 open(newunit=u, file=filename, status='replace', action='write')
 close(u)
 
-call h%initialize(filename, ierr, status='old', action='read')
+call h%open(filename, ierr, status='old', action='read')
 if (ierr==0) error stop 'should have had ierr/=0 on invalid HDF5 file'
 end subroutine test_nonhdf5_file
 
@@ -54,14 +54,14 @@ integer :: u
 type(hdf5_file) :: h
 character(*), parameter :: filename = 'bad.h5'
 
-call h%initialize(filename, status='replace', verbose=.false.)
+call h%open(filename, status='replace', verbose=.false.)
 call h%write('/real32', 42.)
-call h%finalize()
+call h%close()
 
-call h%initialize(filename, status='old', action='read', verbose=.false.)
+call h%open(filename, status='old', action='read', verbose=.false.)
 call h%read('/real32', u)
 if (u /= 42) error stop 'test_wrong_type: did not coerce real to integer'
-call h%finalize()
+call h%close()
 
 end subroutine test_wrong_type
 
