+++ src/tests/test_shape.f90
@@ -11,7 +11,7 @@ integer(HSIZE_T), allocatable :: dims(:)
 
 integer :: d2(3,4), d7(2,1,3,4,7,6,5)
 
-call h%initialize(path, status='scratch')
+call h%open(path, status='scratch')
 call h%write('/d2', d2)
 call h%write('/d7', d7)
 
@@ -23,6 +23,6 @@ call h%shape('/d7', dims)
 if (h%ndims('/d7') /= size(dims)) error stop 'rank /= size(dims)'
 if (any(dims /= shape(d7))) error stop '7-D: file shape not match variable shape'
 
-call h%finalize()
+call h%close()
 
 end program
