### 2021-06-14 6e164f0
ci: add ubuntu 18.04 gcc7 test case
