+++ README.md
@@ -50,8 +50,8 @@ Tested on systems with HDF5 1.8, 1.10 and 1.12 including:
 
 Compilers known to work include:
 
-* Gfortran &ge; 7
-* Intel oneAPI HPC compiler
+* GCC (gfortran) &ge; 7
+* Intel oneAPI HPC compiler &ge; 2021
 
 ---
 
