### 2021-06-17 e4d9956
zlib,hdf5: put args in easier to read variable
doc

update zlib name for Windows, now same as unix
