+++ cmake/build_hdf5.cmake
@@ -11,11 +11,7 @@ include(ExternalProject)
 
 # need to be sure _ROOT isn't empty, defined is not enough
 if(NOT HDF5_ROOT)
-  if(CMAKE_INSTALL_PREFIX_INITIALIZED_TO_DEFAULT)
-    set(HDF5_ROOT ${PROJECT_BINARY_DIR})
-  else()
-    set(HDF5_ROOT ${CMAKE_INSTALL_PREFIX})
-  endif()
+  set(HDF5_ROOT ${CMAKE_INSTALL_PREFIX})
 endif()
 
 set(HDF5_LIBRARIES)
@@ -36,12 +32,26 @@ endif()
 # --- HDF5
 # https://forum.hdfgroup.org/t/issues-when-using-hdf5-as-a-git-submodule-and-using-cmake-with-add-subdirectory/7189/2
 
+set(hdf5_cmake_args
+${zlib_root}
+-DCMAKE_INSTALL_PREFIX:PATH=${HDF5_ROOT}
+-DHDF5_GENERATE_HEADERS:BOOL=false
+-DHDF5_DISABLE_COMPILER_WARNINGS:BOOL=true
+-DBUILD_SHARED_LIBS:BOOL=false
+-DHDF5_ENABLE_PARALLEL:BOOL=false
+-DCMAKE_BUILD_TYPE=Release
+-DHDF5_BUILD_FORTRAN:BOOL=true
+-DHDF5_BUILD_CPP_LIB:BOOL=false
+-DHDF5_BUILD_TOOLS:BOOL=false
+-DBUILD_TESTING:BOOL=false
+-DHDF5_BUILD_EXAMPLES:BOOL=false)
+
 ExternalProject_Add(HDF5
 URL ${hdf5_url}
 URL_HASH SHA256=${hdf5_sha256}
 CONFIGURE_HANDLED_BY_BUILD ON
 INACTIVITY_TIMEOUT 15
-CMAKE_ARGS ${zlib_root} -DCMAKE_INSTALL_PREFIX:PATH=${HDF5_ROOT} -DHDF5_GENERATE_HEADERS:BOOL=false -DHDF5_DISABLE_COMPILER_WARNINGS:BOOL=true -DBUILD_SHARED_LIBS:BOOL=false -DCMAKE_BUILD_TYPE=Release -DHDF5_BUILD_FORTRAN:BOOL=true -DHDF5_BUILD_CPP_LIB:BOOL=false -DHDF5_BUILD_TOOLS:BOOL=false -DBUILD_TESTING:BOOL=false -DHDF5_BUILD_EXAMPLES:BOOL=false
+CMAKE_ARGS ${hdf5_cmake_args}
 BUILD_BYPRODUCTS ${HDF5_LIBRARIES}
 DEPENDS ZLIB)
 
