### 2021-06-17 6d37086
zlib: finesse naming logic
