+++ cmake/build_hdf5.cmake
@@ -38,13 +38,18 @@ ${zlib_root}
 -DHDF5_GENERATE_HEADERS:BOOL=false
 -DHDF5_DISABLE_COMPILER_WARNINGS:BOOL=true
 -DBUILD_SHARED_LIBS:BOOL=false
--DHDF5_ENABLE_PARALLEL:BOOL=false
 -DCMAKE_BUILD_TYPE=Release
 -DHDF5_BUILD_FORTRAN:BOOL=true
 -DHDF5_BUILD_CPP_LIB:BOOL=false
 -DHDF5_BUILD_TOOLS:BOOL=false
 -DBUILD_TESTING:BOOL=false
 -DHDF5_BUILD_EXAMPLES:BOOL=false)
+
+if(hdf5_parallel)
+  list(APPEND hdf5_cmake_args -DHDF5_ENABLE_PARALLEL:BOOL=true)
+else()
+  list(APPEND hdf5_cmake_args -DHDF5_ENABLE_PARALLEL:BOOL=false)
+endif()
 
 ExternalProject_Add(HDF5
 URL ${hdf5_url}
