+++ cmake/Modules/FindHDF5.cmake
@@ -30,6 +30,8 @@ Components
 
 ``Fortran``
 
+``parallel`` checks that the MPI parallel HDF5 layer is enabled
+
 The ``HL`` component is implied and silently accepted to keep
 compatibility with factory FindHDF5
 
@@ -40,6 +42,10 @@ Targets
 ``HDF5::HDF5``
   HDF5 Imported Target
 #]=======================================================================]
+
+include(CheckSymbolExists)
+include(CheckCSourceCompiles)
+include(CheckFortranSourceCompiles)
 
 function(detect_config)
 
@@ -53,14 +59,13 @@ endif()
 
 set(CMAKE_REQUIRED_INCLUDES ${HDF5_INCLUDE_DIR})
 
-foreach(f H5pubconf.h H5pubconf-64.h)
-  if(EXISTS ${HDF5_INCLUDE_DIR}/${f})
-    set(_conf ${HDF5_INCLUDE_DIR}/${f})
-    break()
-  endif()
-endforeach()
+find_file(h5_conf
+  NAMES H5pubconf.h H5pubconf-64.h
+  HINTS ${HDF5_INCLUDE_DIR}
+  NO_DEFAULT_PATH
+)
 
-if(NOT _conf)
+if(NOT h5_conf)
   message(WARNING "Could not find HDF5 config header H5pubconf.h, therefore cannot detect HDF5 configuration.")
   set(HDF5_C_FOUND false PARENT_SCOPE)
   return()
@@ -68,11 +73,30 @@ endif()
 
 # get version
 # from CMake/Modules/FindHDF5.cmake
-include(CheckSymbolExists)
-check_symbol_exists(H5_HAVE_FILTER_SZIP ${_conf} _szip)
-check_symbol_exists(H5_HAVE_FILTER_DEFLATE ${_conf} _zlib)
+check_symbol_exists(H5_HAVE_FILTER_SZIP ${h5_conf} _szip)
+check_symbol_exists(H5_HAVE_FILTER_DEFLATE ${h5_conf} _zlib)
 
-file(STRINGS ${_conf} _def
+if(parallel IN_LIST HDF5_FIND_COMPONENTS)
+  check_symbol_exists(H5_HAVE_PARALLEL ${h5_conf} _h5_parallel)
+  if(_h5_parallel)
+    if(NOT TARGET MPI::MPI_C)
+      find_package(MPI COMPONENTS C)
+    endif()
+    if(MPI_C_FOUND)
+      list(APPEND CMAKE_REQUIRED_LIBRARIES MPI::MPI_C)
+      set(HDF5_parallel_FOUND true PARENT_SCOPE)
+      set(HDF5_IS_PARALLEL true CACHE BOOL "HDF5 library compiled with parallel IO support")
+    endif()
+  endif()
+
+  if(NOT HDF5_IS_PARALLEL)
+    message(WARNING "HDF5 parallel was requested but not found")
+    set(HDF5_parallel_FOUND false PARENT_SCOPE)
+    return()
+  endif()
+endif()
+
+file(STRINGS ${h5_conf} _def
 REGEX "^[ \t]*#[ \t]*define[ \t]+H5_VERSION[ \t]+" )
 if( "${_def}" MATCHES
 "H5_VERSION[ \t]+\"([0-9]+\\.[0-9]+\\.[0-9]+)(-patch([0-9]+))?\"" )
@@ -250,23 +274,16 @@ if(HDF5_Fortran_FOUND)
 
 set(CMAKE_REQUIRED_INCLUDES ${HDF5_Fortran_INCLUDE_DIR} ${HDF5_INCLUDE_DIR})
 
-include(CheckFortranSourceCompiles)
[…]
