+++ .archive/.appveyor.yml
@@ -1,17 +1,0 @@
-image:
-- Ubuntu1804
-
-clone_depth: 25
-
-build: off
-
-init:
-- sudo apt-get install -yq --no-install-suggests --no-install-recommends libhdf5-dev > /dev/null
-
-install:
-- cmake -B build
-- cmake --build build --parallel
-
-test_script:
-- cd build
-- ctest --parallel --output-on-failure
