+++ .archive/.travis.yml
@@ -1,37 +1,0 @@
-language: python
-group: travis_latest
-dist: xenial
-
-git:
-  depth: 25
-  quiet: true
-
-addons:
-  apt:
-    sources: ubuntu-toolchain-r-test
-    packages:
-    - gfortran-6
-    - libhdf5-dev
-
-matrix:
-  include:
-  - os: linux
-    env: FC=gfortran-6
-# OSX OK, but move to Azure for speed
-  # - os: osx
-    # env: FC=gfortran
-    # before_install:
-    # - brew update > /dev/null
-    # - brew upgrade cmake > /dev/null
-    # - brew install gcc > /dev/null
-    # - brew install hdf5 > /dev/null
-
-before_install:
-- git -C $HOME clone https://github.com/mesonbuild/meson
-- python -m pip install -e $HOME/meson
-- curl -Ls https://github.com/ninja-build/ninja/releases/download/v1.9.0/ninja-linux.zip -o ninja-linux.zip
-- sudo unzip ninja-linux.zip -d /usr/local/bin/
-
-install: meson build
-
-script: meson test -C build -v
