### 2021-06-18 d193a05
build_hdf5: build tools and print summary [skip ci]
