+++ cmake/build_hdf5.cmake
@@ -22,7 +22,9 @@ endforeach()
 set(HDF5_INCLUDE_DIRS ${HDF5_ROOT}/include)
 
 # --- Zlib
-set(zlib_root -DHDF5_ENABLE_Z_LIB_SUPPORT:BOOL=ON -DZLIB_USE_EXTERNAL:BOOL=OFF)
+set(zlib_root
+-DHDF5_ENABLE_Z_LIB_SUPPORT:BOOL=ON
+-DZLIB_USE_EXTERNAL:BOOL=OFF)
 
 if(TARGET ZLIB::ZLIB)
   add_custom_target(ZLIB)
@@ -41,7 +43,7 @@ ${zlib_root}
 -DCMAKE_BUILD_TYPE=Release
 -DHDF5_BUILD_FORTRAN:BOOL=true
 -DHDF5_BUILD_CPP_LIB:BOOL=false
--DHDF5_BUILD_TOOLS:BOOL=false
+-DHDF5_BUILD_TOOLS:BOOL=true
 -DBUILD_TESTING:BOOL=false
 -DHDF5_BUILD_EXAMPLES:BOOL=false)
 
