### 2021-06-18 b55e04a
build compiler wrappers on Unix, if building parallel require MPI
