### 2021-06-21 d66f05b
FindHDF5: optimize logic and corner cases
