### 2021-06-23 f95daa5
note -Dautobuild=on default behavior
