### 2021-06-23 #2842 @scivision
The autobuilding of HDF5 and ZLIB is useful:

* even if the system has working HDF5, having the binaries optimized for your CPU and using the latest version of HDF5 and ZLIB can make HDF5 operations up to multiple times faster
* typically HPC have broken or ABI-incompatible HDF5 libraries

However as in f95daa5f there is a `cmake -Dautobuild=off` option to have CMake configure fail if a compatible HDF5 or ZLIB is not found
