### 2021-06-23 52cf545
doc: logical/boolean not yet implemented
