+++ paper/paper.bib
@@ -58,6 +58,13 @@ year = 2019,
 url = {https://github.com/DavidAce/h5pp},
 }
 
+@online{h5xx,
+author= {Felix Höfling},
+title = {h5xx — a template-based C++ wrapper for the HDF5 library}
+year = 2020,
+url = {https://github.com/fhoefling/h5xx},
+}
+
 @article{zettergren,
 author = {Zettergren, M. D. and Snively, J. B.},
 title = {Latitude and Longitude Dependence of Ionospheric TEC and Magnetic Perturbations From Infrasonic-Acoustic Waves Generated by Strong Seismic Events},
