+++ ford.md
@@ -1,9 +1,9 @@
 src_dir: ./src
 output_dir: ./docs
-project: Object-oriented Fortran 2008 HDF5 interface
+project: Object-oriented Fortran HDF5 interface
 project_github: https://github.com/geospace-code/h5fortran
 project_website: https://geospace-code.github.io/h5fortran
-summary: Object-oriented Fortran 2008 HDF5 interface
+summary: Object-oriented Fortran HDF5 interface
 author: Michael Hirsch, Ph.D.
 github: https://github.com/geospace-code
 license: by
@@ -17,3 +17,159 @@ display: public
 source: false
 graph: true
 search: true
+
+
+Simple, robust, thin HDF5 polymorphic Fortran read/write interface.
+Reading or writing {real64,real32,int32,int64} from scalar to 7d is as simple as
+
+```fortran
+use h5fortran
+
+call h5write('my.h5', '/x', x)
+
+call h5read('my.h5', '/y', y)
+```
+
+For NetCDF4 see [nc4fortran](https://github.com/geospace-code/nc4fortran/).
+h5fortran is designed for "serial" HDF5 read/write.
+We don't yet implement the interface for "parallel" HDF5.
+
+h5fortran is designed for easy use using **static** or **shared** linking from your project via:
+
+* `cmake --install`
+* CMake ExternalProject
+* CMake FetchContent
+* CMake + Git submodule
+* Meson subproject
+
+Uses Fortran `submodule` for clean template structure.
+This easy-to-use, thin object-oriented modern Fortran library abstracts away the messy parts of HDF5 so that you can read / write various types/ranks of data with a single command.
+In distinction from other high-level HDF5 interfaces, h5fortran works to deduplicate code, using polymorphism wherever feasible and extensive test suite.
+
+Polymorphic [API](./API.md) with read/write for types int32, real32, real64 with rank scalar (0-D) through 7-D.
+64-bit integers int64 are read/write from scalar through 3-D.
+
+as well as **character (string)**.
+If you need int64, we have a working example for that: src/concepts/int64.f90 that can easily be put into the h5fortran API--just make a GitHub Issue.
+
+* HDF5 **attributes** are also supported for read/write with type character, int32, real32, real64.
+* **Array slicing on read and write** is supported, that is, reading or writing part of a disk HDF5 array into a variable matching the slice shape.
+* Mismatched datatypes are coerced as per standard Fortran rules. For example, reading a float HDF5 variable into an integer Fortran variable:  42.3 => 42
+* Zlib (deflate) compression / decompression -- h5fortran will work without Zlib, but will save/load uncompressed data only.
+* create HDF5 soft link variables--arbitrarily many soft-linked variable names can point to an actual variable, which need not yet exist.
+
+Tested on systems with HDF5 1.8, 1.10 and 1.12 including:
+
+* MacOS (homebrew)
+* Linux (Ubuntu, CentOS)
+* Windows Subsystem for Linux
+* Windows MSYS2
+* IBM Power with Gfortran
+
+Compilers known to work include:
+
+* GCC (gfortran) &ge; 7
+* Intel oneAPI HPC compiler &ge; 2021
+
+---
+
+We welcome [contributions](https://github.com/geospace-code/.github/blob/main/CONTRIBUTING.md).
+In general we hold to the geospace-code [code of conduct](https://github.com/geospace-code/.github/blob/main/CODE_OF_CONDUCT.md).
+
+## Build
+
+Using CMake:
+
+```sh
+git clone https://github.com/geospace-code/h5fortran.git
+
+cd h5fortran
+
+cmake -B build
+
+cmake --build build
+```
[…]
