+++ src/interface.f90
@@ -40,9 +40,11 @@ integer :: libversion(3)  !< major, minor, rel
 
 contains
 !> define methods (procedures) that don't need generic procedure
-procedure, public :: initialize => hdf_initialize, finalize => hdf_finalize, &
+procedure, public :: initialize => hdf_initialize, open => hdf_initialize, &
+  finalize => hdf_finalize, close => hdf_finalize, &
   write_group, create => hdf_create, &
-  open => hdf_open_group, close => hdf_close_group, flush => hdf_flush, &
+  open_group => hdf_open_group, close_group => hdf_close_group, &
+  flush => hdf_flush, &
   ndims => hdf_get_ndims, &
   shape => hdf_get_shape, layout => hdf_get_layout, chunks => hdf_get_chunk, &
   exist => hdf_check_exist, exists => hdf_check_exist, &
@@ -52,8 +54,29 @@ procedure, public :: initialize => hdf_initialize, finalize => hdf_finalize, &
 !> below are procedure that need generic mapping (type or rank agnostic)
 
 !> write group or dataset integer/real
-generic, public :: write => hdf_write_scalar, hdf_write_1d, hdf_write_2d, hdf_write_3d, &
-hdf_write_4d, hdf_write_5d, hdf_write_6d, hdf_write_7d
+generic, public :: write => &
+hdf_write_scalar_r32,hdf_write_scalar_r64,hdf_write_scalar_i32,hdf_write_scalar_i64, hdf_write_scalar_char, &
+hdf_write_1d_r32, hdf_write_1d_r64, hdf_write_1d_i32, hdf_write_1d_i64, &
+hdf_write_2d_r32, hdf_write_2d_r64, hdf_write_2d_i32, hdf_write_2d_i64, &
+hdf_write_3d_r32, hdf_write_3d_r64, hdf_write_3d_i32, hdf_write_3d_i64, &
+hdf_write_4d_r32, hdf_write_4d_r64, hdf_write_4d_i32, hdf_write_4d_i64, &
+hdf_write_5d_r32, hdf_write_5d_r64, hdf_write_5d_i32, hdf_write_5d_i64, &
+hdf_write_6d_r32, hdf_write_6d_r64, hdf_write_6d_i32, hdf_write_6d_i64, &
+hdf_write_7d_r32, hdf_write_7d_r64, hdf_write_7d_i32, hdf_write_7d_i64
+
+generic, public :: write_r32 => hdf_write_scalar_r32,hdf_write_1d_r32,hdf_write_2d_r32,hdf_write_3d_r32, &
+hdf_write_4d_r32,hdf_write_5d_r32,hdf_write_6d_r32,hdf_write_7d_r32
+
+generic, public :: write_r64 => hdf_write_scalar_r64,hdf_write_1d_r64,hdf_write_2d_r64,hdf_write_3d_r64, &
+hdf_write_4d_r64,hdf_write_5d_r64,hdf_write_6d_r64,hdf_write_7d_r64
+
+generic, public :: write_i32 => hdf_write_scalar_i32,hdf_write_1d_i32,hdf_write_2d_i32,hdf_write_3d_i32, &
+hdf_write_4d_i32,hdf_write_5d_i32,hdf_write_6d_i32,hdf_write_7d_i32
+
+generic, public :: write_i64 => hdf_write_scalar_i64,hdf_write_1d_i64,hdf_write_2d_i64,hdf_write_3d_i64, &
+hdf_write_4d_i64,hdf_write_5d_i64,hdf_write_6d_i64,hdf_write_7d_i64
+
+generic, public :: write_char => hdf_write_scalar_char
 
 !> write attributes
[…]
