### 2021-06-23 #2842 @scivision
I  believe I addressed all https://github.com/openjournals/joss-reviews/issues/2842#issuecomment-840427552 in commits just now. Thank you!
