+++ paper/paper.bib
@@ -60,7 +60,7 @@ url = {https://github.com/DavidAce/h5pp},
 
 @online{h5xx,
 author= {Felix Höfling},
-title = {h5xx — a template-based C++ wrapper for the HDF5 library}
+title = {h5xx — a template-based C++ wrapper for the HDF5 library},
 year = 2020,
 url = {https://github.com/fhoefling/h5xx},
 }
