### 2021-06-23 3a18a2d
join citation with proper md citation syntax
