### 2021-06-23 #2842 @pdebuyl
@scivision thank you for the updates

Can you check the version number that is listed in the first comment of this issue page? If it is not up to date, please state what version should be used and provide the zenodo DOI for the archived version of the code.

The metadata of the zenodo archive should have the title "h5fortran: object-oriented polymorphic Fortran interface for HDF5 file IO", only you as "author" and possible other contributors as "contributors" to the entry.
