### 2021-06-23 #2842 @pdebuyl
PS: I also used the joined citation syntax and filed a PR about it.
