### 2021-06-23 #2842 @kyleniemeyer
Hi @scivision, I am doing some final checks before publishing. I noticed two minor issues, could you please resolve theses?

- [ ] In two places (the final paragraph of Summary, and at the end of the Statement of Need) you have references that should use the `[@author1:2001; @author2:2001]` citation syntax, rather than ``[@author1:2001][@author2:2001]` as you have now (this will put them in a single citation separated by commas)
- [ ] In the Collette 2013 reference, please remove the quotation marks around O'Reilly Media, Inc. You might also add the ISBN via the field `isbn = {9781449367831},`
