+++ paper/paper.bib
@@ -32,7 +32,8 @@
   title={Python and HDF5: Unlocking Scientific Data},
   author={Collette, Andrew},
   year={2013},
-  publisher={" O'Reilly Media, Inc."}
+  isbn = {9781449367831},
+  publisher={O'Reilly Media, Inc.}
 }
 
 @online{nc4fortran,
