### 2021-06-25 8d9fa59
FindSZIP: also find libaec
