+++ cmake/Modules/FindSZIP.cmake
@@ -33,11 +33,27 @@ URL "http://www.compressconsult.com/szip/"
 DESCRIPTION "compression library"
 PURPOSE "Some system HDF5 libraries have dynamic links to SZIP")
 
-find_library(SZIP_LIBRARY
-  NAMES szip sz
-  NAMES_PER_DIR)
+find_library(SZIP_AEC_LIBRARY
+  NAMES aec libaec
+  NAMES_PER_DIR
+  DOC "libaec replaces szip with BSD license")
 
-find_path(SZIP_INCLUDE_DIR NAMES szlib.h)
+find_library(SZIP_SZ_LIBRARY
+  NAMES szip sz
+  NAMES_PER_DIR
+  DOC "SZIP API")
+
+set(SZIP_LIBRARY)
+if(SZIP_AEC_LIBRARY)
+  list(APPEND SZIP_LIBRARY ${SZIP_AEC_LIBRARY})
+endif()
+if(SZIP_SZ_LIBRARY)
+  list(APPEND SZIP_LIBRARY ${SZIP_SZ_LIBRARY})
+endif()
+
+find_path(SZIP_INCLUDE_DIR
+  NAMES szlib.h libaec.h
+  DOC "SZIP header")
 
 include(FindPackageHandleStandardArgs)
 find_package_handle_standard_args(SZIP
@@ -48,10 +64,10 @@ if(SZIP_FOUND)
   set(SZIP_LIBRARIES ${SZIP_LIBRARY})
 
   if(NOT TARGET SZIP::SZIP)
-    add_library(SZIP::SZIP UNKNOWN IMPORTED)
+    add_library(SZIP::SZIP INTERFACE IMPORTED)
     set_target_properties(SZIP::SZIP PROPERTIES
-      IMPORTED_LOCATION ${SZIP_LIBRARY}
-      INTERFACE_INCLUDE_DIRECTORIES ${SZIP_INCLUDE_DIR})
+      INTERFACE_LINK_LIBRARIES "${SZIP_LIBRARIES}"
+      INTERFACE_INCLUDE_DIRECTORIES "${SZIP_INCLUDE_DIRS}")
   endif()
 endif()
 
